==> Data analysis/01 - Calculate rolling mean correlations surf bot.Rmd <==
---
title: "Calculate correlations"
author: "Abby Lewis"
date: "2023-05-22"
output: html_document
---

This file calculates the correlation between mean air temperature and surface- and bottom-water temperature and dissolved oxygen.

Table of contents:

- Step 1: Load packages and data
- Step 2: Load and format air temperature data
- Step 3: Calculate and export 30-day rolling mean correlations
- Step 4: Try using Spearman's correlation, rather than partial Spearman's correlation
- Step 5: Re-run with 10-day window
- Step 6: Re-run with 50-day window
- Step 7: Re-run for July 15th to August 31st
- Step 8: Re-run for June, July and August

Step 1: Load packages and data

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
source("correlations_doy.R")

# Load saved data
with_temp = read.csv("../Compiled data/All_data_annual_3m.csv")
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")

# Make a directory for Figures if there isn't one
if (!dir.exists("../Figures")){dir.create("../Figures")}
```

Step 2: Load and format air temperature data

```{r}
daily_temp_raw <- read.csv("../Compiled data/historical_met_output_era5_daily.csv")

daily_temp <- daily_temp_raw %>%
  mutate(Date = as.Date(Date),
         Date_unif = ifelse(Lat < 0, Date + months(6), Date),
         Date_unif = as.Date(Date_unif, origin = "1970-01-01")) %>%
  dplyr::select(-Lat, -Lon) %>%
  mutate(Year = year(Date)) %>%
  filter(Year <= 2022)

rm(daily_temp_raw)
```

Step 3: Calculate and export rolling means (30-day)

```{r}
many_lake_temp_sum <- with_temp %>%
  left_join(daily_temp) %>%
  mutate(doy = yday(Date_unif),
         Year = ifelse(month(Date_unif) >= 9, Year + 1, Year)) %>%
  group_by(LakeID) %>%
  arrange(Date_unif) %>%
  mutate(Temp_C = zoo::rollmean(Temp_C, 30, align = "right", fill = NA))

# ---

hypo_do_roll <- many_lake_temp_sum %>%
  mutate(hypoxic = ifelse(median(DO_mgL_BOT, na.rm = T) < 1,"yes","no")) %>%
  filter(hypoxic=="no") %>%
  correlations_doy("DO_mgL_BOT", "Temp_C", "hypolimnetic DO")

hypo_demand_roll <- correlations_doy(many_lake_temp_sum, 
                                    "DO_demand_mgLd", "Temp_C", 
                                    "Summer VHOD ")

epi_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_SURF", "Temp_C", 
                                    "Summer epi. temperature ")

hypo_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_BOT", "Temp_C", 
                                    "Summer hypo. temperature ")

hypo_do_sat_roll <- many_lake_temp_sum %>%
  mutate(hypoxic = ifelse(median(DO_mgL_BOT, na.rm = T) < 1,"yes","no")) %>%
  filter(hypoxic=="no") %>%
  correlations_doy("DO_sat_BOT", "Temp_C", "Summer hypo. DO sat ")

all <- hypo_do_roll %>%
  mutate(var = "Summer hypo. DO") %>%
  bind_rows(epi_temp_roll %>%
              mutate(var = "Summer epi. temperature")) %>%
  bind_rows(hypo_temp_roll %>%
              mutate(var = "Summer hypo. temperature")) %>%
  bind_rows(hypo_do_sat_roll %>%
              mutate(var = "Summer hypo. DO sat")) %>%
  bind_rows(hypo_demand_roll %>% 
              mutate(var = "Summer VHOD")) 

write.csv(all, 
          "../Compiled data/Correlations - 30 day rolling mean - surf bot.csv", 
          row.names = F)

all %>%
  group_by(LakeID, var) %>%
  summarize(nyear = unique(nyear)) %>%
  group_by(var) %>%
  summarize(min = min(nyear),
            max = max(nyear),
            median = median(nyear))
```

Step 4: Try using Spearman's correlation, rather than partial Spearman's correlation

```{r}
hypo_do_roll <- many_lake_temp_sum %>%
  mutate(hypoxic = ifelse(median(DO_mgL_BOT, na.rm = T) < 1,"yes","no")) %>%
  filter(hypoxic=="no") %>%
  correlations_doy("DO_mgL_BOT", "Temp_C", "hypolimnetic DO", partial = F)

hypo_demand_roll <- correlations_doy(many_lake_temp_sum, 
                                    "DO_demand_mgLd", "Temp_C", 
                                    "Summer VHOD ", partial = F)

epi_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_SURF", "Temp_C", 
                                    "Summer epi. temperature ", partial = F)

hypo_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_BOT", "Temp_C", 
                                    "Summer hypo. temperature ", partial = F)

all <- hypo_do_roll %>%
  mutate(var = "Summer hypo. DO") %>%
  bind_rows(epi_temp_roll %>%
              mutate(var = "Summer epi. temperature")) %>%
  bind_rows(hypo_temp_roll %>%
              mutate(var = "Summer hypo. temperature")) %>%
  bind_rows(hypo_demand_roll %>% 
              mutate(var = "Summer VHOD")) 

write.csv(all, "../Compiled data/Correlations - 30 day rolling mean - not partial surf bot.csv", row.names = F)
```

Step 5: Re-run with 10-day window

```{r}
many_lake_temp_sum <- with_temp %>%
  left_join(daily_temp) %>%
  mutate(doy = yday(Date_unif),
         Year = ifelse(month(Date_unif) >= 9, Year + 1, Year)) %>%
  group_by(LakeID) %>%
  arrange(Date_unif) %>%
  mutate(Temp_C = zoo::rollmean(Temp_C, 10, align = "right", fill = NA))

# ---

hypo_do_roll <- many_lake_temp_sum %>%
  mutate(hypoxic = ifelse(median(DO_mgL_BOT, na.rm = T) < 1,"yes","no")) %>%
  filter(hypoxic=="no") %>%
  correlations_doy("DO_mgL_BOT", "Temp_C", "hypolimnetic DO")

hypo_demand_roll <- correlations_doy(many_lake_temp_sum, 
                                    "DO_demand_mgLd", "Temp_C", 
                                    "Summer VHOD ")

epi_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_SURF", "Temp_C", 
                                    "Summer epi. temperature ")

hypo_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_BOT", "Temp_C", 
                                    "Summer hypo. temperature ")

all <- hypo_do_roll %>%
  mutate(var = "Summer hypo. DO") %>%
  bind_rows(epi_temp_roll %>%
              mutate(var = "Summer epi. temperature")) %>%
  bind_rows(hypo_temp_roll %>%
              mutate(var = "Summer hypo. temperature")) %>%
  bind_rows(hypo_demand_roll %>% 
              mutate(var = "Summer VHOD")) 

write.csv(all, "../Compiled data/Correlations - 10 day rolling mean surf bot.csv", row.names = F)
```

Step 6: Re-run with 50-day window

```{r}
many_lake_temp_sum <- with_temp %>%
  left_join(daily_temp) %>%
  mutate(doy = yday(Date_unif),
         Year = ifelse(month(Date_unif) >= 9, Year + 1, Year)) %>%
  group_by(LakeID) %>%
  arrange(Date_unif) %>%
  mutate(Temp_C = zoo::rollmean(Temp_C, 50, align = "right", fill = NA))

# ---

hypo_do_roll <- many_lake_temp_sum %>%
  mutate(hypoxic = ifelse(median(DO_mgL_BOT, na.rm = T) < 1,"yes","no")) %>%
  filter(hypoxic=="no") %>%
  correlations_doy("DO_mgL_BOT", "Temp_C", "hypolimnetic DO")

hypo_demand_roll <- correlations_doy(many_lake_temp_sum, 
                                    "DO_demand_mgLd", "Temp_C", 
                                    "Summer VHOD ")

epi_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_SURF", "Temp_C", 
                                    "Summer epi. temperature ")

hypo_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_BOT", "Temp_C", 
                                    "Summer hypo. temperature ")

all <- hypo_do_roll %>%
  mutate(var = "Summer hypo. DO") %>%
  bind_rows(epi_temp_roll %>%
              mutate(var = "Summer epi. temperature")) %>%
  bind_rows(hypo_temp_roll %>%
              mutate(var = "Summer hypo. temperature")) %>%
  bind_rows(hypo_demand_roll %>% 
              mutate(var = "Summer VHOD")) 

write.csv(all, "../Compiled data/Correlations - 50 day rolling mean surf bot.csv", row.names = F)
```

Step 7: Re-run for July 15th to August 31st

```{r}
with_temp <- read.csv("../Compiled data/All_data_annual_15th_to_31st.csv")

many_lake_temp_sum <- with_temp %>%
  left_join(daily_temp) %>%
  mutate(doy = yday(Date_unif),
         Year = ifelse(month(Date_unif) >= 9, Year + 1, Year)) %>%
  group_by(LakeID) %>%
  arrange(Date_unif) %>%
  mutate(Temp_C = zoo::rollmean(Temp_C, 30, align = "right", fill = NA))

# ---

hypo_do_roll <- many_lake_temp_sum %>%
  mutate(hypoxic = ifelse(median(DO_mgL_BOT, na.rm = T) < 1,"yes","no")) %>%
  filter(hypoxic=="no") %>%
  correlations_doy("DO_mgL_BOT", "Temp_C", "hypolimnetic DO")

epi_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_SURF", "Temp_C", 
                                    "Summer epi. temperature ")

hypo_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_BOT", "Temp_C", 
                                    "Summer hypo. temperature ")

all <- hypo_do_roll %>%
  mutate(var = "Summer hypo. DO") %>%
  bind_rows(epi_temp_roll %>%
              mutate(var = "Summer epi. temperature")) %>%
  bind_rows(hypo_temp_roll %>%
              mutate(var = "Summer hypo. temperature"))

write.csv(all, "../Compiled data/Correlations - 30 day rolling mean - 15th to 31st surf bot.csv", row.names = F)
```

Step 8: Re-run for June, July and August

```{r}
with_temp <- read.csv("../Compiled data/All_data_annual_JJA.csv")

many_lake_temp_sum <- with_temp %>%
  left_join(daily_temp) %>%
  mutate(doy = yday(Date_unif),
         Year = ifelse(month(Date_unif) >= 9, Year + 1, Year)) %>%
  group_by(LakeID) %>%
  arrange(Date_unif) %>%
  mutate(Temp_C = zoo::rollmean(Temp_C, 30, align = "right", fill = NA))

# ---

hypo_do_roll <- many_lake_temp_sum %>%
  mutate(hypoxic = ifelse(median(DO_mgL_BOT, na.rm = T) < 1,"yes","no")) %>%
  filter(hypoxic=="no") %>%
  correlations_doy("DO_mgL_BOT", "Temp_C", "hypolimnetic DO")

epi_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_SURF", "Temp_C", 
                                    "Summer epi. temperature ")

hypo_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_BOT", "Temp_C", 
                                    "Summer hypo. temperature ")

all <- hypo_do_roll %>%
  mutate(var = "Summer hypo. DO") %>%
  bind_rows(epi_temp_roll %>%
              mutate(var = "Summer epi. temperature")) %>%
  bind_rows(hypo_temp_roll %>%
              mutate(var = "Summer hypo. temperature"))

write.csv(all, "../Compiled data/Correlations - 30 day rolling mean - JJA surf bot.csv", row.names = F)
```
