==> Data analysis/02 - Data characterization.Rmd <==
---
title: "Data characterization"
author: "Abby Lewis"
date: "2023-05-22"
output: html_document
---

This code file performs basic characterization of the dataset and generates a map of study sites

Table of contents:

- Step 1: Load packages
- Step 2: Summary statistics
- Step 3: Map distribution of lakes
- Step 4: Generate data summary plots


Step 1: Load packages

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(viridis)
library(lubridate)

# Make a directory for Figures if there isn't one
if (!dir.exists("../Figures")){dir.create("../Figures")}
```

Step 2: Summary statistics

```{r}
#Load data from EDI
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579") 
#Load memory calculations to  identify which lakes were included in the analysis
mem <- read.csv("../Compiled data/Correlations - 30 day rolling mean - surf bot.csv")

for_summary <- mem %>%
  left_join(lat_long)

#Number of lakes
length(unique(for_summary$LakeID))

#Distribution
length(unique(for_summary$Country))
unique(for_summary$Country)

#Lake size
for_summary %>%
  dplyr::reframe(across(.cols = c(MaximumDepth_m, 
                                  SurfaceArea_ha, 
                                  Elevation_m, 
                                  Latitude_DD),
                          .fns = ~ c(median(.x,na.rm=T), 
                                     min(.x,na.rm=T), 
                                     max(.x,na.rm=T))))

#Duration
annual <- read_csv("../Compiled data/All_data_annual_3m.csv") %>%
  filter(LakeID %in% for_summary$LakeID) 

annual %>%
  group_by(LakeID) %>%
  dplyr::summarize(duration = as.numeric(max(Year)-min(Year))+1) %>%
  ungroup() %>%
  dplyr::summarize(min_duration = min(duration, na.rm = T),
                   max_duration = max(duration, na.rm = T),
                   sd_duration = sd(duration, na.rm = T),
                   mean_duration = mean(duration, na.rm =T),
                   median_duration = median(duration, na.rm = T))

#Years per variable
annual %>%
  filter(LakeID %in% for_summary$LakeID) %>%
  group_by(LakeID) %>%
  mutate(medianDO = median(DO_mgL_BOT, na.rm = T),
         DO_mgL_BOT = ifelse(medianDO < 1, NA, DO_mgL_BOT)) %>% #Only when median DO > 1
  select(DO_mgL_BOT, Temp_C_SURF, Temp_C_BOT, DO_demand_mgLd, LakeID, Year) %>%
  pivot_longer(DO_mgL_BOT:DO_demand_mgLd) %>%
  filter(!is.na(value)) %>%
  group_by(name, LakeID) %>%
  dplyr::summarize(years = length(unique(Year))) %>%
  filter(years >= 15) %>%
  group_by(name) %>%
  dplyr::summarize(min = min(years, na.rm = T),
                   max = max(years, na.rm = T),
                   median = median(years, na.rm = T))


lat_long_for_sum <- lat_long %>%
  filter(LakeID %in% mem$LakeID) %>%
  filter(abs(Latitude_DD) >= 23.5)

#Daily
all_daily <- read_csv("../Compiled data/surf_bot_daily_all.csv")
daily <- read_csv("../Compiled data/surf_bot_daily.csv")
daily %>%
  filter(LakeID %in% mem$LakeID) %>%
  group_by(LakeID) %>%
  mutate(medianDO = median(DO_mgL_BOT, na.rm = T),
         DO_mgL_BOT = ifelse(medianDO < 1, NA, DO_mgL_BOT)) %>% #Only when median DO > 1
  select(DO_mgL_BOT, Temp_C_SURF, Temp_C_BOT, LakeID, Date_unif) %>%
  pivot_longer(DO_mgL_BOT:Temp_C_BOT) %>%
  filter(!is.na(value)) %>%
  mutate(Year = year(Date_unif)) %>%
  group_by(name, LakeID) %>%
  filter(length(unique(Year)) >= 15) %>%
  group_by(name, LakeID, Year) %>%
  dplyr::summarize(days = length(unique(Date_unif))) %>%
  group_by(name, LakeID) %>%
  dplyr::summarize(days = median(days)) %>%
  group_by(name) %>%
  dplyr::summarize(min = min(days, na.rm = T),
                   max = max(days, na.rm = T),
                   median = median(days, na.rm = T))

#Demand
demand <- read_csv("../Compiled data/VW oxygen demand points.csv")
demand %>%
  filter(LakeID %in% mem$LakeID) %>%
  group_by(LakeID) %>%
  filter(length(unique(Year))>= 15) %>%
  ungroup() %>%
  summarize(median = median(n),
            max = max(n),
            min = min(n))
```

Step 3: Map distribution of lakes

```{r}
library(ggthemes)
library(sf)
library(rnaturalearth)
library(rnaturalearthdata)
#remotes::install_github("ropensci/rnaturalearthhires")
library(rnaturalearthhires)
library(ggspatial)

#Create a world map of the sites in this analysis
world <- ne_countries(scale = "medium", returnclass = "sf")
states <- ne_states(returnclass = "sf",country = "United States of America")
for_map <- lat_long_for_sum

jpeg("../Figures/World Map - just points.jpg", res = 300, width = 6, height = 3.2, units = "in")
world_map <- ggplot(data = world) +
  geom_sf(fill = "white", color = "grey60") +
  coord_sf(expand = FALSE, ylim = c(-50,80))+
  geom_point(data = for_map, aes(Longitude_DD, Latitude_DD), 
             fill = "#3087D9",shape = 21, color = "white", size = 2, alpha  =.7, stroke = .4)+
  theme_bw()+
  theme(plot.margin = ggplot2::margin(0, 0, 0, 0, "cm"),
    legend.box.background = element_rect(fill = "white", color = "white"))+
  scale_fill_viridis(name = "Years of data")+
  theme(axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        panel.background = element_rect(fill = "grey93"),
        panel.grid = element_blank())
world_map
dev.off()
```

Step 4: Generate data summary plots

```{r}
depth <- lat_long_for_sum %>%
  ggplot(aes(x = MaximumDepth_m))+
  geom_density(color = "#3087D9")+
  scale_x_log10(breaks = c(3, 30, 300))+
  theme_bw(base_size = 9)+
  xlab(expression("Maximum depth (m)"))+
  theme(plot.margin = ggplot2::margin(5.5, 8, 5.5, 8, "pt"))

area <- lat_long_for_sum %>%
  ggplot(aes(x = SurfaceArea_ha/100))+
  geom_density(color = "#3087D9")+
  scale_x_log10(breaks = c(0.1, 10, 1000),
                labels = c(0.1, 10, 1000))+
  theme_bw(base_size = 9)+
  xlab(expression("Surface area (km"^{2}*")"))+
  theme(plot.margin = ggplot2::margin(5.5, 8, 5.5, 8, "pt"),
        axis.title.x=element_text(vjust=2))

years <- annual %>%
  group_by(Year) %>%
  summarize(lakes = length(unique(LakeID))) %>% 
  ggplot(aes(x = Year, y = lakes))+
  geom_line(color = "#3087D9")+
  ylab("Lakes (n)")+
  xlab(expression("Year"))+
  theme_bw(base_size = 9)+
  scale_x_continuous(n.breaks = 3.5)+
  theme(plot.margin = ggplot2::margin(5.5, 8, 5.5, 8, "pt"))

doys <- all_daily %>%
  left_join(lat_long %>%
              select(LakeID, Latitude_DD)) %>%
  mutate(Date_unif = as.Date(Date_unif)) %>%
  mutate(yday = lubridate::yday(Date_unif)) %>%
  group_by(LakeID, yday) %>%
  summarize() %>%
  ggplot(aes(x = yday))+
  geom_density(color = "#3087D9")+
  xlab(expression("Sampling date"))+
  theme_bw(base_size = 9)+
  scale_x_continuous(n.breaks = 3,
                     breaks = c(yday("2024-06-20"), yday("2024-12-21")),
                     labels = c("Summer\nsolstice", "Winter\nsolstice"))+
  theme(plot.margin = ggplot2::margin(5.5, 8, 5.5, 8, "pt"))

jpeg("../Figures/Figure 2.jpeg", width = 6, height = 4, units = "in", res = 600)
ggpubr::ggarrange(world_map, 
                  ggpubr::ggarrange(depth, area, doys, years, nrow = 1, ncol = 4, align = "h", 
                                    labels = c("b1", "b2", "b3", "b4", "b5"), label.y = 1.04,
                                    font.label = list(size = 10)), 
                  nrow=2, heights = c(1, .5), labels = c("a", "b"), label.y = c(1.03, 1.2), label.x = c(0.05, 0.05))
dev.off()
```
