==> Data analysis/04 - Calculate memory and plot rolling means surf bot.Rmd <==
---
title: "Air temp analysis"
author: "Abby Lewis"
date: "2023-05-22"
output: html_document
editor_options: 
  markdown: 
    wrap: 72
---

This file calculates "seasonal ecological memory" as the maximum absolute correlation between our summer focal variables and 30-day rolling mean air temperature during July--August (summer) minus the maximum correlation during spring (March-May).

Table of contents:

-   Step 1: Load packages and data
-   Step 2: Calculate seasonal ecological memory
-   Step 3: Generate Figure 1 (summary boxplot)
-   Step 4: Compare hypo DO concentration (mg/L) to hypo DO saturation (%)
-   Step 5: Plot rolling means (hypolimnetic variables)
-   Step 6: Export dates
-   Step 7: Map
-   Step 8: Plot multiple rolling-mean windows (SI)
-   Step 9: Plot multiple summer definitions (SI)
-   Step 10: Plot partial and not-partial correlations
-   Step 11: Try median instead of max
-   Step 12: Try different definitions of spring

Step 1: Load packages and data

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)

#Set global options
sp <- yday(c("2022-03-01", "2022-05-31"))
su <- yday(c("2022-07-01", "2022-08-31"))
alpha <- 0.001

#Load saved data
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")
all <- read.csv("../Compiled data/Correlations - 30 day rolling mean - surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F)) #note that doy is unif (NH and SH)
onset <- read.csv("../Compiled data/stratification_onset.csv")

# Make a directory for Figures if there isn't one
if (!dir.exists("../Figures")){dir.create("../Figures")}

#Labs for plots
levels_no_n = c("Summer epi. temperature", 
                "Summer hypo. temperature",
                "Summer VHOD",
                "Summer hypo. DO")

labels_no_n = c("Surface-water\ntemperature",
                "Bottom-water\ntemperature",
                "Bottom-water\noxygen demand",
                "Bottom-water\ndissolved oxygen")

stratified <- read_csv("../Compiled data/stratified_lakes.csv")
```

Step 2: Calculate seasonal ecological memory

```{r}
memory <- all %>% 
  filter(nyear >= 15,
         var %in% levels_no_n) %>%
  mutate(var = factor(var, 
                      levels = levels_no_n,
                      labels = labels_no_n),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season) %>%
  summarize(min = abs(monthly_correlation[which.min(monthly_correlation)]),
            max = abs(monthly_correlation[which.max(monthly_correlation)])) %>%
  ungroup() %>%
  mutate(val = ifelse(var %in% c("Bottom-water\ndissolved oxygen"),
                         min, max)) %>%
  select(-min, -max) %>%
  pivot_wider(names_from = season, values_from = val) %>%
  mutate(M = (spring - summer)) %>%
  group_by(var) %>%
  mutate(p = wilcox.test(M)$p.value)
```

Step 3: Generate Figure 3 (summary boxplot)

```{r}
jpeg("../Figures/Figure 3.jpg", width = 3.5, height = 3, res = 600, units = "in")
memory %>%
  filter(!var == "Summer\nepi. DO sat") %>%
  group_by(var) %>%
  mutate(n = length(unique(LakeID)),
         var = paste0(var, " (n = ",n,")")) %>%
  mutate(var = factor(var, 
                      levels = c("Surface-water\ntemperature (n = 365)",
                                 "Bottom-water\ntemperature (n = 365)",
                                 "Bottom-water\noxygen demand (n = 157)",
                                 "Bottom-water\ndissolved oxygen (n = 213)"),
                      labels = c("Surface-water\ntemperature (n = 365)",
                                 "Bottom-water\ntemperature (n = 365)",
                                 "Bottom-water\noxygen demand (n = 157)",
                                 "Bottom-water\ndissolved oxygen (n = 213)")))%>%
  group_by(var) %>%
  mutate(color = ifelse(p > alpha, "Insignificant", 
                        ifelse(median(M) > 0, "median > 0", "median < 0")),
         color = factor(color, levels = c("median > 0", 
                                          "Insignificant", 
                                          "median < 0"))) %>%
  ggplot(aes(y = var, x = M, color = color))+
  geom_vline(xintercept = 0, color = "grey70")+
  geom_boxplot()+
  annotate("text", x = max(memory$M) + 0.1, y = 5, label = "More correlated\nwith spring", 
           lineheight = .9, size = 3, hjust = 0, color = "#25469a")+
  annotate("text", x = min(memory$M) - 0.1, y = 5, label = "More correlated\nwith summer", 
           lineheight = .9, size = 3, hjust = 1, color = "#c16586")+
  coord_cartesian(xlim = c(max(memory$M), min(memory$M)),
                  ylim = c(1,4), clip = "off")+
  theme_bw() +
  ggtitle("\n")+
  scale_y_discrete(limits=rev, position = "right") +
  xlab("Relative correlation with spring\nvs. summer air temperatures")+
  guides(color = guide_legend(nrow = 2))+
  theme(axis.title.y = element_blank(),
        axis.title.x = element_text(size = 8),
        legend.position = "none")+
  scale_color_manual(values = c("#25469a","#c16586", "black"), 
                     name = paste0("Boxplot color\n(p < ", alpha, ")"))+
  scale_x_reverse()
dev.off()
```

Step 4: Compare hypo DO concentration (mg/L) to hypo DO saturation (%)

```{r}
memory_do <- all %>% 
  filter(nyear >= 15) %>%
  filter(var %in% c("Summer hypo. DO",
                    "Summer hypo. DO sat")) %>%
  mutate(var = factor(var, 
                      levels = c("Summer hypo. DO",
                                 "Summer hypo. DO sat"),
                      labels = c("Bottom-water\nDO (mg/L)",
                                 "Bottom-water\nDO (% saturation)")),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season) %>%
  summarize(min = abs(monthly_correlation[which.min(monthly_correlation)]),
            max = abs(monthly_correlation[which.max(monthly_correlation)])) %>%
  ungroup() %>%
  mutate(val = min) %>%
  select(-min, -max) %>%
  pivot_wider(names_from = season, values_from = val) %>%
  mutate(M = (spring - summer)) %>%
  group_by(var) %>%
  mutate(p = wilcox.test(M)$p.value,
         n = length(unique(LakeID)),
         var = paste0(var, "\nn = ",n)) %>%
  group_by(var) %>%
  mutate(color = ifelse(p > alpha, "Insignificant", 
                        ifelse(median(M) > 0, "median > 0", "median < 0")),
         color = factor(color, levels = c("median < 0", 
                                          "Insignificant", 
                                          "median > 0"))) 

jpeg("../Figures/Monthly correlations - DO sat comp surf bot.jpg", width = 3.5, height = 2.5, res = 600, units = "in")
memory_do %>%
  ggplot(aes(y = var, x = M, color = color))+
  geom_vline(xintercept = 0, color = "grey70")+
  geom_boxplot()+
  annotate("text", x = max(memory_do$M), y = 3, label = "More correlated\nwith spring", 
           lineheight = .9, size = 3, hjust = 0, color = "#25469a")+
  annotate("text", x = min(memory_do$M), y = 3, label = "More correlated\nwith summer", 
           lineheight = .9, size = 3, hjust = 1, color = "#c16586")+
  theme_bw() +
  ggtitle("\n")+
  scale_y_discrete(limits=rev, position = "right") +
  xlab("Relative correlation with summer\nvs. spring air temperature")+
  guides(color = guide_legend(nrow = 2))+
  theme(axis.title.y = element_blank(),
        axis.title.x = element_text(size= 8),
        legend.position = "none",
        legend.text = element_text(size = 8),
        legend.key.size = unit(1, "line"),
        legend.title = element_text(size = 8),
        legend.margin=ggplot2::margin(c(0,0,0,0)),
        legend.box.margin=ggplot2::margin(t = -2,l = 0, r = 0, b = -5),
        #axis.text.y=element_text(colour=c("#25469a","#25469a","#25469a","#c16586"))
        )+
  scale_color_manual(values = c("#25469a","#c16586"), 
                     name = paste0("Boxplot color\n(p < ", alpha, ")"))+
  coord_cartesian(xlim = c(max(memory_do$M),min(memory_do$M)), 
                  ylim = c(1,2), clip = "off") +
  scale_x_reverse()
dev.off()
```

Step 5: Export dates

```{r}
dates_to_plot <- all %>%
  filter(nyear >= 15) %>%
  mutate(season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  filter(!is.na(season)) %>%
  group_by(var, LakeID, season) %>%
  summarize(max_doy = doy[which.max(abs(monthly_correlation))]) 

write.csv(dates_to_plot, 
          "../Compiled data/strongest_correlation_dates_all_met_3m.csv", 
          row.names = F)
```

Step 6: Plot rolling means (hypolimnetic variables)

```{r}
#Function to deliniate periods of significant vs insignificant correlations
group_fun <- function(sig) {
  group <- rep(1, length(sig))
  for (i in 2:length(sig)) {
    if(sig[i] == sig[i - 1]) {
      group[i] <- group[i - 1]
    } else {
      group[i] <- group[i - 1] + 1
    }
  }
  return(group)
}

#Calculate mean and SD
all_plot <- all %>%
  filter(nyear >= 15) %>%
  mutate(strat = LakeID %in% stratified$LakeID) %>%
  group_by(var, strat) %>%
  mutate(n = length(unique(LakeID)), #Some lakes are missing from January because our air temp only goes back to 1980
         var = paste0(var, " (n = ",n,")")) %>%
  group_by(var, doy, strat) %>%
    mutate(p = wilcox.test(monthly_correlation)$p.value,
           sig = p < 0.001) %>%
  group_by(var, doy, sig, strat) %>%
  summarise(mean = mean(monthly_correlation, na.rm = T),
            sd = sd(monthly_correlation, na.rm = T)) 

#Use group function
all_plot$group <- 1
for(i in 1:length(unique(all_plot$var))){
  all_plot$group[all_plot$var == unique(all_plot$var)[i]] <- 
    group_fun(all_plot$sig[all_plot$var == unique(all_plot$var)[i]])
}

#To get smooth ribbons, we need to manually duplicate datapoints
all_plot1 <- all_plot %>%
  mutate(DateTime = as.Date("2022-01-01") + days(doy))
all_plot2 <- all_plot %>%
  ungroup() %>%
  mutate(DateTime = as.Date("2022-01-01") + days(doy) + hours(23) + minutes(59),
         mean = ifelse(!is.na(lead(var)) & lead(var) == var, lead(mean), mean),
         sd = ifelse(lead(var) == var, lead(sd), sd))

dates_to_plot_strat <- dates_to_plot %>%
  filter(LakeID %in% stratified$LakeID) %>%
  mutate(var = factor(var, 
                      levels = c("Summer hypo. temperature",
                                 "Summer VHOD",
                                 "Summer hypo. DO"
                                 ),
                      labels = c("Summer bottom-water\ntemperature (n = 336)",
                                 "Summer bottom-water\noxygen demand (n = 155)",
                                 "Summer bottom-water\nDO (n = 178)")),
         x_val = as.Date("2022-01-01")+days(max_doy)) %>%
  filter(!is.na(var))

dates_to_plot_unstrat <- dates_to_plot %>%
  filter(!LakeID %in% stratified$LakeID) %>%
  mutate(var = factor(var, 
                      levels = c("Summer hypo. temperature",
                                 "Summer hypo. DO"
                                 ),
                      labels = c("Summer bottom-water\ntemperature (n = 29)",
                                 "Summer bottom-water\nDO (n = 35)")),
         x_val = as.Date("2022-01-01")+days(max_doy)) %>%
  filter(!is.na(var))

#Plot hypolimnetic variables
hypo_vars_strat <-all_plot1 %>%
  bind_rows(all_plot2) %>%
  filter(doy>=30) %>% #remove first 29 days (these include previous year) 
  filter(strat == T) %>%
  filter(var %in% c("Summer VHOD (n = 155)",
                    #"Summer epi. temperature (n = 317)",
                    "Summer hypo. DO (n = 178)",
                    "Summer hypo. temperature (n = 336)")) %>%
  mutate(var = factor(var, 
                      levels = c(#"Summer epi. temperature (n = 469)",
                                 "Summer hypo. temperature (n = 336)",
                                 "Summer VHOD (n = 155)",
                                 "Summer hypo. DO (n = 178)"
                                 ),
                      labels = c(#"Summer surface-water\ntemperature (n = 469)",
                                 "Summer bottom-water\ntemperature (n = 336)",
                                 "Summer bottom-water\noxygen demand (n = 155)",
                                 "Summer bottom-water\nDO (n = 178)")),
         layer = ifelse(grepl("epi", var), "epi", "hypo")) %>%
  ggplot() + 
  geom_hline(yintercept=0, color = "grey60", linewidth = 0.3)+
  geom_vline(xintercept = as_datetime("2022-01-01")+
               days(median(onset$Onset, na.rm = T)),
             color = "grey50", lty = "dashed")+
  geom_ribbon(aes(x = DateTime, 
                  ymin = mean - sd, ymax = mean + sd,
                  fill = sig, group = group),
              alpha = 0.5) +
  #geom_jitter(aes(x = as_datetime(x_val), 
  #               y = 0.6,
  #               color = season), 
  #           shape = 21, size = 0.5, width = 0, height = 0.1,
  #           data = dates_to_plot_strat)+
  geom_line(aes(x = DateTime, y = mean)) +
  xlab("Air temperature date (right-aligned 30-day rolling mean)")+
  ylab("Correlation")+
  theme_bw()+
  scale_fill_manual(values = c("grey40","#0FA9E6")) +
  labs(fill = paste0("p < ",alpha)) +
  facet_wrap(~var, nrow = 1, drop = F) +
  theme(legend.position = "none",
        axis.title.x = element_blank(),
        strip.background = element_blank()) +
  ggtitle("Dimictic and warm-monomictic lakes")

hypo_vars_unstrat <- all_plot1 %>%
  bind_rows(all_plot2) %>%
  filter(doy>=30) %>% #remove first 29 days (these include previous year)
  filter(var %in% c(#"Summer epi. temperature (n = 92)",
                    "Summer hypo. DO (n = 35)",
                    "Summer hypo. temperature (n = 29)")) %>%
  mutate(var = factor(var, 
                      levels = c(#"Summer epi. temperature (n = 92)",
                                 "Summer hypo. temperature (n = 29)",
                                 "Summer VHOD (n = 2)",
                                 "Summer hypo. DO (n = 35)"
                                 ),
                      labels = c(#"Summer surface-water\ntemperature (n = 92)",
                                 "Summer bottom-water\ntemperature (n = 29)",
                                 "Summer bottom-water\noxygen demand",
                                 "Summer bottom-water\nDO (n = 35)")),
         layer = ifelse(grepl("epi", var), "epi", "hypo")) %>%
  filter(!var == "Summer bottom-water\noxygen demand") %>%
  ggplot() + 
  geom_hline(yintercept=0, color = "grey60", linewidth = 0.3)+
  geom_ribbon(aes(x = DateTime, 
                  ymin = mean - sd, ymax = mean + sd,
                  fill = sig, group = group),
              alpha = 0.5) +
  #geom_jitter(aes(x = as_datetime(x_val), 
  #               y = 0.6,
  #               color = season), 
  #           shape = 21, size = 0.5, width = 0, height = 0.1,
  #           data = dates_to_plot_unstrat)+
  geom_line(aes(x = DateTime, y = mean)) +
  xlab("Air temperature date (right-aligned 30-day rolling mean)")+
  ylab("Correlation")+
  theme_bw()+
  scale_fill_manual(values = c("grey40","#0FA9E6"))+
  labs(fill = paste0("p < ",alpha)) +
  facet_wrap(~var, nrow = 1, drop = F) +
  theme(legend.position = "bottom",
        legend.margin=ggplot2::margin(c(0,0,0,0)),
        strip.background = element_blank())+
  ggtitle("Polymictic lakes")

#Save plot
jpeg("../Figures/Monthly correlations - hypo daily roll mean sd surf bot.jpg", 
     width = 7, height = 4, res = 300, units = "in")
ggpubr::ggarrange(hypo_vars_strat, hypo_vars_unstrat, 
                  nrow = 2, align = "hv", common.legend = T, legend = "bottom")
dev.off()
```

Step 7: Map

```{r}
library(ggthemes)
library(sf)
library(rnaturalearth)
library(rnaturalearthdata)
#remotes::install_github("ropensci/rnaturalearthhires")
library(rnaturalearthhires)
library(ggspatial)
library(viridis)

#Create a world map of the sites in this analysis
world <- ne_countries(scale = "medium", returnclass = "sf")
states <- ne_states(returnclass = "sf",country = "United States of America")
lat_lon <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")

for_map <- memory %>%
  filter(var == "Bottom-water\ntemperature") %>%
  left_join(lat_lon)

jpeg("../Figures/World Map - memory surf bot.jpg", res = 300, width = 6, height = 3.2, units = "in")
world_map <- ggplot(data = world) +
  geom_sf(fill = "white") +
  coord_sf(expand = FALSE)+
  geom_point(data = for_map, aes(Longitude_DD, Latitude_DD, fill = M), 
             shape = 21, color = "grey50", size = 2, alpha  =.7, stroke = .4)+
  theme_bw()+
  scale_fill_gradient2(name = "CSEM", high = "#25469a", low = "#c16586")+
  theme(plot.margin = ggplot2::margin(0, 0, 0, 0, "cm"),
        legend.box.background = element_rect(fill = "white", color = "white"),
        axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        axis.ticks.x = element_blank(),
        axis.text.x = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        panel.background = element_rect(fill = "grey90"),
        panel.grid = element_line(color = "grey80",linetype = "dotted", linewidth = 0.0001))
world_map
dev.off()

for_states <- for_map%>%
  filter(Country == "USA")
us_map <- ggplot(data = states) +
  geom_sf(data = world, fill = "grey98") +
  geom_sf(fill = "white") +
  coord_sf(expand = FALSE, ylim = c(25, 50), xlim = c(-130,-60))+
  geom_point(data = for_map, aes(Longitude_DD, Latitude_DD, fill = M), 
             shape = 21, color = "grey50", size = 2, alpha  =.7, stroke = .4)+
  theme_bw()+
  scale_fill_gradient2(name = "CSEM", high = "#25469a", low = "#c16586")+
  annotation_north_arrow(location = "bl", which_north = "true", 
                         pad_x = unit(0.2, "in"), pad_y = unit(0.2, "in"),
                         style = north_arrow_fancy_orienteering) +
  theme(plot.margin = ggplot2::margin(0, 0, 0, 0, "cm"),
        legend.box.background = element_rect(fill = "white", color = "white"),
        panel.background = element_rect(fill = "grey90"),
        panel.grid = element_blank(),
        axis.title.x = element_blank(),
        axis.title.y = element_blank())
jpeg("../Figures/US Map - memory surf bot.jpg", res = 300, width = 6, height = 3.2, units = "in")
us_map
dev.off()

euro_map <- ggplot(data = world) +
  geom_sf(fill = "white") +
  coord_sf(expand = FALSE, ylim = c(35, 70), xlim = c(-12,40))+
  geom_point(data = for_map, aes(Longitude_DD, Latitude_DD, fill = M), 
             shape = 21, color = "grey50", size = 2, alpha  =.7, stroke = .4)+
  theme_bw()+
  scale_fill_gradient2(name = "CSEM", high = "#25469a", low = "#c16586")+
  annotation_north_arrow(location = "bl", which_north = "true", 
      pad_x = unit(0.2, "in"), pad_y = unit(0.2, "in"),
      style = north_arrow_fancy_orienteering) +
  theme(plot.margin = ggplot2::margin(0, 0, 0, 0, "cm"),
    legend.box.background = element_rect(fill = "white", color = "white")) +
  theme(plot.margin = ggplot2::margin(0, 0, 0, 0, "cm"),
        legend.box.background = element_rect(fill = "white", color = "white"),
        panel.background = element_rect(fill = "grey90"),
        panel.grid = element_blank(),
        axis.title.x = element_blank(),
        axis.title.y = element_blank())
jpeg("../Figures/Euro Map - memory surf bot.jpg", res = 300, width = 6, height = 3.2, units = "in")
euro_map
dev.off()

for_wi <- for_map#%>%
  #filter(Country == "USA", 
  #       StateOrProvince == "Wisconsin")
wi_map <- ggplot(data = states) +
  geom_sf(data = world, fill = "grey90") +
  geom_sf(fill = "white") +
  coord_sf(expand = FALSE, ylim = c(42, 47.5), xlim = c(-93,-86.5))+
  geom_point(data = for_map, aes(Longitude_DD, Latitude_DD, fill = M), 
             shape = 21, color = "grey50", size = 2, alpha  =.7, stroke = .4)+
  theme_bw()+
  scale_fill_gradient2(name = "CSEM", high = "#25469a", low = "#c16586")+
  #annotation_north_arrow(location = "bl", which_north = "true", 
  #    pad_x = unit(0.2, "in"), pad_y = unit(0.2, "in"),
  #    style = north_arrow_fancy_orienteering) +
  theme(plot.margin = ggplot2::margin(0, 0, 0, 0, "cm"),
    legend.box.background = element_rect(fill = "white", color = "white"))+
  scale_x_continuous(breaks = seq(-93, -87, 2))+
  theme(axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        panel.background = element_rect(fill = "grey90"),
        panel.grid = element_blank())
wi_map

for_states <- for_map#%>%
  #filter(Country == "USA")
states <- ne_states(returnclass = "sf",country = "United States of America")
ne_map <- ggplot(data = states) +
  geom_sf(data = world, fill = "grey98") +
  geom_sf(fill = "white") +
  coord_sf(expand = FALSE, ylim = c(42, 47.5), xlim = c(-76,-67))+
  geom_point(data = for_map, aes(Longitude_DD, Latitude_DD, fill = M), 
             shape = 21, color = "grey50", size = 2, alpha  =.7, stroke = .4)+
  theme_bw()+
  scale_fill_gradient2(name = "CSEM", high = "#25469a", low = "#c16586")+
  #annotation_north_arrow(location = "bl", which_north = "true", 
  #    pad_x = unit(0.2, "in"), pad_y = unit(0.2, "in"),
  #    style = north_arrow_fancy_orienteering) +
  theme(plot.margin = ggplot2::margin(0, 0, 0, 0, "cm"),
    legend.box.background = element_rect(fill = "white", color = "white"))+
  theme(axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        panel.background = element_rect(fill = "grey90"),
        panel.grid = element_blank())
ne_map

library(ggpubr)
library(patchwork)
jpeg("../Figures/World map with insets surf bot.jpg", res = 300, width = 6, height = 5, units = "in")
ggarrange(world_map + theme(legend.position = "none"), 
          ggarrange(wi_map + theme(legend.position = "none"), 
                    ne_map, 
                    plot_spacer()+theme_void(), 
                    labels = c("b", "c", NA), widths = c(4,7,1), nrow = 1), 
          nrow = 2, heights = c(2,1), labels = c("a",NA), vjust = 1)
dev.off()
```

Step 8: Plot multiple rolling-mean windows (SI)

```{r}
alpha <- 1
all_30 <- read.csv("../Compiled data/Correlations - 30 day rolling mean - surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F),
         window = "30-day rolling mean")
all_50 <- read.csv("../Compiled data/Correlations - 50 day rolling mean surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F),
         window = "50-day rolling mean")
all_10 <- read.csv("../Compiled data/Correlations - 10 day rolling mean surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F),
         window = "10-day rolling mean")

all <- all_30 %>%
  full_join(all_50) %>%
  full_join(all_10)

memory_windows <- all %>% 
  filter(nyear >= 15) %>%
  filter(var %in% c("Summer epi. temperature",
                    "Summer epi. DO sat",
                    "Summer hypo. temperature",
                    "Summer VHOD",
                    "Summer hypo. DO")) %>%
  mutate(var = factor(var, levels = levels_no_n,
                      labels = labels_no_n),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season, window) %>%
  summarize(min = abs(monthly_correlation[which.min(monthly_correlation)]),
            max = abs(monthly_correlation[which.max(monthly_correlation)])) %>%
  ungroup() %>%
  mutate(val = ifelse(var %in% c("Bottom-water\ndissolved oxygen"),
                         min, max)) %>%
  select(-min, -max) %>%
  pivot_wider(names_from = season, values_from = val) %>%
  mutate(M = (spring - summer)) %>%
  group_by(var, window) %>%
  mutate(p = wilcox.test(M)$p.value)

jpeg("../Figures/Monthly correlations - sum long concept v2 many windows.jpg", 
     width = 7, height = 3, res = 600, units = "in")
memory_windows %>%
  filter(!var == "Summer\nepi. DO sat") %>%
  group_by(var, window) %>%
  mutate(var = factor(var, 
                      levels = labels_no_n,
                      labels = labels_no_n))%>%
  group_by(var, window) %>%
  mutate(color = ifelse(p > alpha, "Insignificant", 
                        ifelse(median(M) > 0, "median > 0", "median < 0")),
         color = factor(color, levels = c("median < 0", 
                                          "Insignificant", 
                                          "median > 0"))) %>%
  ggplot(aes(y = var, x = M, color = color))+
  geom_vline(xintercept = 0, color = "grey70")+
  geom_boxplot()+
  annotate("text", x = max(memory_windows$M), y = 5.8, label = "More correlated\nwith spring", 
           lineheight = .9, size = 3, hjust = 0, color = "#25469a")+
  annotate("text", x = min(memory_windows$M), y = 5.8, label = "More correlated\nwith summer", 
           lineheight = .9, size = 3, hjust = 1, color = "#c16586")+
  theme_bw() +
  ggtitle("\n")+
  scale_y_discrete(limits=rev, position = "right") +
  xlab("Relative correlation with summer\nvs. spring air temperature")+
  guides(color = guide_legend(nrow = 2))+
  theme(axis.title.y = element_blank(),
        axis.title.x = element_text(size= 8),
        legend.position = "none",
        legend.text = element_text(size = 8),
        legend.key.size = unit(1, "line"),
        legend.title = element_text(size = 8),
        legend.margin=ggplot2::margin(c(0,0,0,0)),
        legend.box.margin=ggplot2::margin(t = -2,l = 0, r = 0, b = -5),
        #axis.text.y=element_text(colour=c("#25469a","#25469a","#25469a","#c16586"))
        )+
  scale_color_manual(values = c("#c16586","#25469a"), 
                     name = paste0("Boxplot color\n(p < ", alpha, ")"))+
  coord_cartesian(xlim = c(max(memory_windows$M),min(memory_windows$M)), 
                  ylim = c(1,4), clip = "off")+
  facet_wrap(~window)+
  scale_x_reverse()
dev.off()
```

Step 9: Plot multiple summer definitions (SI)

```{r}
all_ja <- read.csv("../Compiled data/Correlations - 30 day rolling mean - surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F),
         window = "July 1 - August 31")
all_jja <- read.csv("../Compiled data/Correlations - 30 day rolling mean - JJA surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F),
         window = "June 1 - August 31")
all_15 <- read.csv("../Compiled data/Correlations - 30 day rolling mean - 15th to 31st surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F),
         window = "July 15 - August 31")

all <- all_ja %>%
  full_join(all_jja) %>%
  full_join(all_15)

memory_mult <- all %>% 
  filter(var %in% c("Summer epi. temperature",
                    "Summer hypo. temperature",
                    "Summer hypo. DO")) %>%
  mutate(var = factor(var, levels = c("Summer epi. temperature",
                                      "Summer hypo. temperature",
                                      "Summer hypo. DO"),
                      labels = c("Surface-water\ntemperature",
                                 "Bottom-water\ntemperature",
                                 "Bottom-water\ndissolved oxygen")),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season, window) %>%
  summarize(min = abs(monthly_correlation[which.min(monthly_correlation)]),
            max = abs(monthly_correlation[which.max(monthly_correlation)])) %>%
  ungroup() %>%
  mutate(val = ifelse(var %in% c("Bottom-water\ndissolved oxygen"),
                         min, max)) %>%
  select(-min, -max) %>%
  pivot_wider(names_from = season, values_from = val) %>%
  mutate(M = (spring - summer)) %>%
  group_by(var, window) %>%
  mutate(p = wilcox.test(M)$p.value)

jpeg("../Figures/Monthly correlations - sum long concept v2 many summers.jpg", 
     width = 7, height = 3, res = 600, units = "in")
memory_mult %>%
  group_by(var, window) %>%
  mutate(color = ifelse(p > alpha, "Insignificant", 
                        ifelse(median(M) > 0, "median > 0", "median < 0")),
         color = factor(color, levels = c("median < 0", 
                                          "Insignificant", 
                                          "median > 0"))) %>%
  ggplot(aes(y = var, x = M, color = color))+
  geom_vline(xintercept = 0, color = "grey70")+
  geom_boxplot()+
  annotate("text", x = max(memory_mult$M), y = 4.4, label = "More correlated\nwith spring", 
           lineheight = .9, size = 3, hjust = 0, color = "#25469a")+
  annotate("text", x = min(memory_mult$M), y = 4.4, label = "More correlated\nwith summer", 
           lineheight = .9, size = 3, hjust = 1, color = "#c16586")+
  theme_bw() +
  ggtitle("\n")+
  scale_y_discrete(limits = rev, position = "right") +
  xlab("Relative correlation with summer\nvs. spring air temperature")+
  guides(color = guide_legend(nrow = 2))+
  theme(axis.title.y = element_blank(),
        axis.title.x = element_text(size = 8),
        legend.position = "none",
        legend.text = element_text(size = 8),
        legend.key.size = unit(1, "line"),
        legend.title = element_text(size = 8),
        legend.margin=ggplot2::margin(c(0,0,0,0)),
        legend.box.margin=ggplot2::margin(t = -2, l = 0, r = 0, b = -5),
        #axis.text.y=element_text(colour=c("#25469a","#25469a","#25469a","#c16586"))
        )+
  scale_color_manual(values = c("#c16586", "#25469a"), 
                     name = paste0("Boxplot color\n(p < ", alpha, ")"))+
  coord_cartesian(xlim = c(max(memory_mult$M), min(memory_mult$M)), 
                  ylim = c(1, 3), clip = "off")+
  facet_wrap(~window)+
  scale_x_reverse()
dev.off()
```

Step 10: Plot partial and not-partial correlations

```{r}
all_partial <- read.csv("../Compiled data/Correlations - 30 day rolling mean - surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F),
         window = "Partial")
all_not_partial <- read.csv("../Compiled data/Correlations - 30 day rolling mean - not partial surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F),
         window = "Not-partial")

all <- all_partial %>%
  full_join(all_not_partial)

memory_partial <- all %>% 
  filter(var %in% levels_no_n) %>%
  mutate(var = factor(var, levels = levels_no_n,
                      labels = labels_no_n),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season, window) %>%
  summarize(min = abs(monthly_correlation[which.min(monthly_correlation)]),
            max = abs(monthly_correlation[which.max(monthly_correlation)])) %>%
  ungroup() %>%
  mutate(val = ifelse(var %in% c("Bottom-water\ndissolved oxygen"),
                         min, max)) %>%
  select(-min, -max) %>%
  pivot_wider(names_from = season, values_from = val) %>%
  mutate(M = (spring - summer)) %>%
  group_by(var, window) %>%
  mutate(p = wilcox.test(M)$p.value)

jpeg("../Figures/Monthly correlations - sum long concept v2 partial not partial.jpg", 
     width = 5, height = 3, res = 600, units = "in")
memory_partial %>%
  group_by(var, window) %>%
  mutate(color = ifelse(p > alpha, "Insignificant", 
                        ifelse(median(M) > 0, "median > 0", "median < 0")),
         color = factor(color, levels = c("median < 0", 
                                          "Insignificant", 
                                          "median > 0"))) %>%
  ggplot(aes(y = var, x = M, color = color))+
  geom_vline(xintercept = 0, color = "grey70")+
  geom_boxplot()+
  annotate("text", x = max(memory_partial$M), y = 5.8, label = "More correlated\nwith spring", 
           lineheight = .9, size = 3, hjust = 0, color = "#25469a")+
  annotate("text", x = min(memory_partial$M), y = 5.8, label = "More correlated\nwith summer", 
           lineheight = .9, size = 3, hjust = 1, color = "#c16586")+
  theme_bw() +
  ggtitle("\n")+
  scale_y_discrete(limits = rev, position = "right") +
  xlab("Relative correlation with summer\nvs. spring air temperature")+
  guides(color = guide_legend(nrow = 2))+
  theme(axis.title.y = element_blank(),
        axis.title.x = element_text(size = 8),
        legend.position = "none",
        legend.text = element_text(size = 8),
        legend.key.size = unit(1, "line"),
        legend.title = element_text(size = 8),
        legend.margin=ggplot2::margin(c(0,0,0,0)),
        legend.box.margin=ggplot2::margin(t = -2, l = 0, r = 0, b = -5),
        #axis.text.y=element_text(colour=c("#25469a","#25469a","#25469a","#c16586"))
        )+
  scale_color_manual(values = c("#c16586", "#25469a"), 
                     name = paste0("Boxplot color\n(p < ", alpha, ")"))+
  coord_cartesian(xlim = c(max(memory_partial$M), min(memory_partial$M)), 
                  ylim = c(1, 4), clip = "off")+
  facet_wrap(~window)+
  scale_x_reverse()
dev.off()
```

Step 11: Try median instead of max

```{r}
all <- read.csv("../Compiled data/Correlations - 30 day rolling mean - surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F)) #note that doy is unif (NH and SH)

memory_mean <- all %>% 
  filter(var %in% levels_no_n) %>%
  mutate(var = factor(var, levels = levels_no_n,
                      labels = labels_no_n),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season) %>%
  summarize(min_val = abs(monthly_correlation[which.min(monthly_correlation)]),
            max_val = abs(monthly_correlation[which.max(monthly_correlation)]),
            median = abs(median(monthly_correlation))) %>%
  ungroup() %>%
  mutate(maximum = ifelse(var %in% c("Bottom-water\ndissolved oxygen"),
                         min_val, max_val)) %>%
  select(-min_val, -max_val) %>%
  pivot_longer(cols = c(median, maximum), names_to = "metric", values_to = "value") %>%
  pivot_wider(names_from = season, values_from = value) %>%
  mutate(M = (spring - summer)) %>%
  group_by(var, metric) %>%
  mutate(p = wilcox.test(M)$p.value)

jpeg("../Figures/Monthly correlations (mean) - sum long concept v2.jpg", 
     width = 5, height = 3, res = 600, units = "in")
memory_mean %>%
  filter(!var == "Summer\nepi. DO sat") %>%
  group_by(var, metric) %>%
  mutate(color = ifelse(p > alpha, "Insignificant", 
                        ifelse(median(M) > 0, "median > 0", "median < 0")),
         color = factor(color, levels = c("median < 0", 
                                          "Insignificant", 
                                          "median > 0"))) %>%
  ggplot(aes(y = var, x = M, color = color))+
  geom_vline(xintercept = 0, color = "grey70")+
  geom_boxplot()+
  annotate("text", x = max(memory_mean$M), y = 5.8, label = "More correlated\nwith spring", 
           lineheight = .9, size = 3, hjust = 0, color = "#25469a")+
  annotate("text", x = min(memory_mean$M), y = 5.8, label = "More correlated\nwith summer", 
           lineheight = .9, size = 3, hjust = 1, color = "#c16586")+
  theme_bw() +
  ggtitle("\n")+
  scale_y_discrete(limits=rev, position = "right") +
  xlab("Relative correlation with summer\nvs. spring air temperature")+
  guides(color = guide_legend(nrow = 2))+
  theme(axis.title.y = element_blank(),
        axis.title.x = element_text(size= 8),
        legend.position = "none",
        legend.text = element_text(size = 8),
        legend.key.size = unit(1, "line"),
        legend.title = element_text(size = 8),
        legend.margin=ggplot2::margin(c(0,0,0,0)),
        legend.box.margin=ggplot2::margin(t = -2,l = 0, r = 0, b = -5),
        #axis.text.y=element_text(colour=c("#25469a","#25469a","#25469a","#c16586"))
        )+
  scale_color_manual(values = c("#c16586", "#25469a"), 
                     name = paste0("Boxplot color\n(p < ", alpha, ")"))+
  coord_cartesian(xlim = c(max(memory_mean$M), min(memory_mean$M)), 
                  ylim = c(1,4), clip = "off") +
  facet_wrap(~metric)+
  scale_x_reverse()
dev.off()
```

Step 12: Try different definitions of spring

```{r}
sp <- yday(c("2022-03-01", "2022-05-01"))
su <- yday(c("2022-07-01", "2022-08-31"))

all <- read.csv("../Compiled data/Correlations - 30 day rolling mean - surf bot.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F)) #note that doy is unif (NH and SH)

memory_short_spring <- all %>% 
  filter(nyear >= 15,
         var %in% levels_no_n) %>%
  mutate(var = factor(var, 
                      levels = levels_no_n,
                      labels = labels_no_n),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season) %>%
  summarize(min = abs(monthly_correlation[which.min(monthly_correlation)]),
            max = abs(monthly_correlation[which.max(monthly_correlation)])) %>%
  ungroup() %>%
  mutate(val = ifelse(var %in% c("Bottom-water\ndissolved oxygen"),
                         min, max)) %>%
  select(-min, -max) %>%
  pivot_wider(names_from = season, values_from = val) %>%
  mutate(M = (spring - summer)) %>%
  group_by(var) %>%
  mutate(p = wilcox.test(M)$p.value,
         duration = "Spring = March 1 - May 1")

sp <- yday(c("2022-03-01", "2022-05-31"))
su <- yday(c("2022-07-01", "2022-08-31"))

memory_long_spring <- all %>% 
  filter(nyear >= 15,
         var %in% levels_no_n) %>%
  mutate(var = factor(var, 
                      levels = levels_no_n,
                      labels = labels_no_n),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season) %>%
  summarize(min = abs(monthly_correlation[which.min(monthly_correlation)]),
            max = abs(monthly_correlation[which.max(monthly_correlation)])) %>%
  ungroup() %>%
  mutate(val = ifelse(var %in% c("Bottom-water\ndissolved oxygen"),
                         min, max)) %>%
  select(-min, -max) %>%
  pivot_wider(names_from = season, values_from = val) %>%
  mutate(M = (spring - summer)) %>%
  group_by(var) %>%
  mutate(p = wilcox.test(M)$p.value,
         duration = "Spring = March 1 - May 31")

jpeg("../Figures/Monthly correlations - changing spring duration.jpg", 
     width = 6, height = 3, res = 600, units = "in")
memory_short_spring %>%
  bind_rows(memory_long_spring) %>%
  filter(!var == "Summer\nepi. DO sat") %>%
  group_by(var, duration) %>%
  mutate(color = ifelse(p > alpha, "Insignificant", 
                        ifelse(median(M) > 0, "median > 0", "median < 0")),
         color = factor(color, levels = c("median < 0", 
                                          "Insignificant", 
                                          "median > 0"))) %>%
  ggplot(aes(y = var, x = M, color = color))+
  geom_vline(xintercept = 0, color = "grey70")+
  geom_boxplot()+
  annotate("text", x = max(memory_mean$M), y = 5.8, label = "More correlated\nwith spring", 
           lineheight = .9, size = 3, hjust = 0, color = "#25469a")+
  annotate("text", x = min(memory_mean$M), y = 5.8, label = "More correlated\nwith summer", 
           lineheight = .9, size = 3, hjust = 1, color = "#c16586")+
  theme_bw() +
  ggtitle("\n")+
  scale_y_discrete(limits=rev, position = "right") +
  xlab("Relative correlation with summer\nvs. spring air temperature")+
  guides(color = guide_legend(nrow = 2))+
  theme(axis.title.y = element_blank(),
        axis.title.x = element_text(size= 8),
        legend.position = "none",
        legend.text = element_text(size = 8),
        legend.key.size = unit(1, "line"),
        legend.title = element_text(size = 8),
        legend.margin=ggplot2::margin(c(0,0,0,0)),
        legend.box.margin=ggplot2::margin(t = -2,l = 0, r = 0, b = -5),
        #axis.text.y=element_text(colour=c("#25469a","#25469a","#25469a","#c16586"))
        )+
  scale_color_manual(values = c("#c16586", "#25469a"), 
                     name = paste0("Boxplot color\n(p < ", alpha, ")"))+
  coord_cartesian(xlim = c(max(memory_mean$M),min(memory_mean$M)), ylim = c(1,4), clip = "off") +
  facet_wrap(~duration)+
  scale_x_reverse()
dev.off()
```
