==> Data analysis/05 - Memory driver analysis - RF - surf bot v2.Rmd <==
---
title: "Air temp analysis"
author: "Abby Lewis"
date: "2023-05-22"
output: html_document
---

This file performs a driver analysis to characterize the extent to which lake morphometric factors explain variation in the strength of seasonal ecological memory

Table of contents: 
- Step 1: Load packages and data 
- Step 2: Calculate seasonal ecological memory
- Step 3: Run driver analysis

Step 1: Load packages and data

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(ggpubr)
library(ggtext)
library(randomForest)
source("run_rf_analysis.R")

#Set global options
sp <- yday(c("2022-03-01", "2022-05-31"))
su <- yday(c("2022-07-01", "2022-08-31"))

#Load saved data
with_temp = read.csv("../Compiled data/All_data_annual_3m.csv")
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")
all <- read.csv("../Compiled data/Correlations - 30 day rolling mean - surf bot.csv")

all %>%
  left_join(lat_long) %>%
  filter(doy>35,
         nyear >= 15,
         abs(Latitude_DD) >= 23.5) %>%
  group_by(LakeID, var) %>%
  summarize(nyear = DescTools::Mode(nyear)) %>%
  group_by(var) %>%
  summarize(min = min(nyear),
            max = max(nyear),
            median = median(nyear))

#Buoyancy frequency
other_drivers <- with_temp %>%
  group_by(LakeID) %>%
  mutate(range_DO = max(DO_mgL_BOT, na.rm = T) - min(DO_mgL_BOT, na.rm = T)) %>%
  select(TP_ugL_SURF, DOC_mgL_SURF, Chla_ugL_SURF, DO_mgL_BOT, range_DO, LakeID, buoyancy_freq) %>%
  summarize(across(everything(), \(x) median(x, na.rm = TRUE)))

# Make a directory for Figures if there isn't one
if (!dir.exists("../Figures")){dir.create("../Figures")}
```

Step 2: Calculate seasonal ecological memory

```{r}
memory <- all %>% 
  filter(nyear >= 15) %>%
  filter(var %in% c("Summer epi. temperature",
                    "Summer hypo. temperature",
                    "Summer VHOD",
                    "Summer hypo. DO")) %>%
  mutate(var = factor(var, levels = c("Summer epi. temperature",
                                      "Summer hypo. temperature",
                                      "Summer VHOD",
                                      "Summer hypo. DO"),
                      labels = c("Surface-water temperature",
                                 "Bottom-water temperature", 
                                 "Bottom-water oxygen demand",
                                 "Bottom-water dissolved oxygen")),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  group_by(var) %>%
  mutate(p = wilcox.test(monthly_correlation)$p.value) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season, p) %>%
  summarize(nyear = max(nyear),
            min = abs(monthly_correlation[which.min(monthly_correlation)]),
            max = abs(monthly_correlation[which.max(monthly_correlation)])) %>%
  ungroup() %>%
  mutate(val = ifelse(var %in% c("Bottom-water dissolved oxygen"),
                         min, max)) %>%
  select(-min, -max) %>%
  pivot_wider(names_from = season, values_from = val) %>%
  mutate(M = (spring - summer)) 

write.csv(memory, "../Compiled data/memory_3m.csv")

for_one_to_one <- memory %>%
  left_join(other_drivers) %>%
  filter(DO_mgL_BOT >= 1 | !var == "Bottom-water dissolved oxygen")

medians <- for_one_to_one %>%
  group_by(var) %>%
  mutate(n = length(unique(LakeID))) %>%
  mutate(layer = ifelse(grepl("Surf", var), "Surface-water", "Bottom-water")) %>%
  group_by(var, layer) %>%
  summarize(spring = median(spring, na.rm = T),
            summer = median(summer, na.rm = T),
            color = summer > spring)

ns <- for_one_to_one %>%
  group_by(var) %>%
  mutate(n = length(unique(LakeID))) %>%
  mutate(layer = ifelse(grepl("Surf", var), "Surface-water", "Bottom-water"),
         color = summer > spring) %>%
  group_by(var, color) %>%
  summarize(n = n(), 
            text = paste0("n = ", n)) %>%
  mutate(y = ifelse(color, .99, -0.1),
         x = ifelse(color, 0, .2),
         text = ifelse(color, paste0("Summer-driven, ", text), paste0("Spring-driven, ", text)))

one_to_one <- for_one_to_one %>%
  group_by(var) %>%
  mutate(layer = ifelse(grepl("Surf", var), "Surface-water", "Bottom-water"),
         color = summer > spring)%>%
  ggplot(aes(x = spring, y = summer, color = color)) +
  geom_abline(slope = 1, color = "grey50") +
  geom_point() +
  facet_wrap(~var) +
  scale_color_manual(values = c("#25469a", "#c16586"))+
  scale_fill_manual(values = c("#25469a", "#c16586"))+
  xlab("Correlation with spring air temperature") +
  ylab("Correlation with summer air temeprature") +
  theme_bw() +
  ylim(-0.1,1.05)+
  theme(legend.position = "none") +
  geom_point(data = medians, color = "white", fill = "black", size = 7, shape = 21, stroke = 0.5) +
  geom_point(data = medians, aes(fill = color), color = "white", size = 4, shape = 21, stroke = 1) +
  geom_text(aes(x = x, y = y, label = text), data = ns, hjust = 0, vjust = 0) 

jpeg("../Figures/one_to_one.jpg", width = 5, height = 5, units = "in", res = 300)
one_to_one
dev.off()
```

Step 3: Run driver analysis

```{r}
mem_rf <- memory %>%
  select(LakeID, var, M, nyear) %>%
  left_join(lat_long) %>%
  left_join(other_drivers) %>%
  mutate(log_SA = log(SurfaceArea_ha),
         log_max_depth = log(MaximumDepth_m),
         log_TP = log(TP_ugL_SURF),
         log_DOC = log(DOC_mgL_SURF),
         log_chla = log(Chla_ugL_SURF),
         abs_lat = abs(Latitude_DD)) %>%
  pivot_wider(names_from = var, values_from = M) %>%
  rename(epi_temp = `Surface-water temperature`,
         hypo_temp = `Bottom-water temperature`,
         hypo_do = `Bottom-water dissolved oxygen`,
         vhod = `Bottom-water oxygen demand`) %>%
  ungroup()

jpeg("../Figures/driver correlations.jpg", width = 5, height = 5, 
     units = "in", res = 400)
GGally::ggpairs(mem_rf %>% 
                  select(c("log_max_depth", "buoyancy_freq", "abs_lat")) %>%
                  rename("Log maximum depth (m)" = log_max_depth,
                         "Buoyancy frequency (1/s)" = buoyancy_freq,
                         "Absolute latitude (DD)" = abs_lat))+
  theme_bw()+
  theme(text = element_text(size = 9))
dev.off()

potential_drivers <- c("buoyancy_freq", "abs_lat", "log_max_depth")
#Figure 4
jpeg("../Figures/Figure 4.jpg", width = 7, height = 4.5, 
       units = "in", res = 600)
run_rf_analysis(potential_drivers, mem_rf)
dev.off()

#20 years
jpeg("../Figures/partials_20year.jpg", width = 7, height = 4.5, 
       units = "in", res = 400)
run_rf_analysis(potential_drivers, mem_rf %>% filter(nyear >= 20))
dev.off()

#25 years
jpeg("../Figures/partials_25year.jpg", width = 7, height = 4.5, 
       units = "in", res = 400)
run_rf_analysis(potential_drivers, mem_rf %>% filter(nyear >= 25))
dev.off()

#Spring
jpeg("../Figures/partials_spring.jpg", width = 7, height = 4.5, 
       units = "in", res = 400)
run_rf_analysis(potential_drivers, 
                memory %>%
                  select(LakeID, var, spring, nyear) %>%
                  left_join(lat_long) %>%
                  left_join(other_drivers) %>%
                  mutate(log_SA = log(SurfaceArea_ha),
                         log_max_depth = log(MaximumDepth_m),
                         log_TP = log(TP_ugL_SURF),
                         log_DOC = log(DOC_mgL_SURF),
                         log_chla = log(Chla_ugL_SURF),
                         abs_lat = abs(Latitude_DD)) %>%
                  pivot_wider(names_from = var, values_from = spring) %>%
                  rename(epi_temp = `Surface-water temperature`,
                         hypo_temp = `Bottom-water temperature`,
                         hypo_do = `Bottom-water dissolved oxygen`,
                         vhod = `Bottom-water oxygen demand`) %>%
                  ungroup(),
                color_vals = "black",
                color_limits = c(-1, 1),
                ylab_text = "Partial dependence of correlation\nwith spring air temperature")
dev.off()

#Summer
jpeg("../Figures/partials_summer.jpg", width = 7, height = 4.5, 
       units = "in", res = 400)
run_rf_analysis(potential_drivers, 
                memory %>%
                  select(LakeID, var, summer, nyear) %>%
                  left_join(lat_long) %>%
                  left_join(other_drivers) %>%
                  mutate(log_SA = log(SurfaceArea_ha),
                         log_max_depth = log(MaximumDepth_m),
                         log_TP = log(TP_ugL_SURF),
                         log_DOC = log(DOC_mgL_SURF),
                         log_chla = log(Chla_ugL_SURF),
                         abs_lat = abs(Latitude_DD)) %>%
                  pivot_wider(names_from = var, values_from = summer) %>%
                  rename(epi_temp = `Surface-water temperature`,
                         hypo_temp = `Bottom-water temperature`,
                         hypo_do = `Bottom-water dissolved oxygen`,
                         vhod = `Bottom-water oxygen demand`) %>%
                  ungroup(),
                color_vals = "black",
                color_limits = c(-1, 1),
                ylab_text = "Partial dependence of correlation\nwith summer air temperature")
dev.off()
```
