==> Data analysis/05 - Memory driver analysis - RF - surf bot.Rmd <==
---
title: "Air temp analysis"
author: "Abby Lewis"
date: "2023-05-22"
output: html_document
---

This file performs a driver analysis to characterize the extent to which lake morphometric factors explain variation in the strength of seasonal ecological memory

Table of contents: 
- Step 1: Load packages and data 
- Step 2: Calculate seasonal ecological memory
- Step 3: Run driver analysis
- Step 4: Plot identified drivers

Step 1: Load packages and data

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(ggpubr)
library(ggtext)
library(randomForest)
source("lmer_functions.R")

#Set global options
sp <- yday(c("2022-03-01", "2022-05-31"))
su <- yday(c("2022-07-01", "2022-08-31"))

#Load saved data
with_temp = read.csv("../Compiled data/All_data_annual_3m.csv")
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")
all <- read.csv("../Compiled data/Correlations - 30 day rolling mean - surf bot.csv") 
#Buoyancy frequency
bf = read.csv("../Compiled data/summer_averages_wi.csv")%>%
  group_by(LakeID) %>%
  summarize(buoyancy_freq = median(buoyancy_freq, na.rm = T))
other_drivers <- with_temp %>%
  group_by(LakeID) %>%
  select(TP_ugL_SURF, DOC_mgL_SURF, Chla_ugL_SURF, LakeID) %>%
  summarize(across(everything(), \(x) mean(x, na.rm = TRUE)))

# Make a directory for Figures if there isn't one
if (!dir.exists("../Figures")){dir.create("../Figures")}
```

Step 2: Calculate seasonal ecological memory

```{r}
memory <- all %>% 
  filter(var %in% c("Summer epi. temperature",
                    "Summer hypo. temperature",
                    "Summer VHOD",
                    "Summer hypo. DO")) %>%
  mutate(var = factor(var, levels = c("Summer epi. temperature",
                                      "Summer hypo. temperature",
                                      "Summer VHOD",
                                      "Summer hypo. DO"),
                      labels = c("Summer epi. temperature",
                                 "Summer bottom-water\ntemperature",
                                 "Summer VHOD",
                                 "Summer bottom-water\ndissolved oxygen")),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  group_by(var) %>%
  mutate(p = wilcox.test(monthly_correlation)$p.value) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season, p) %>%
  summarize(max = max(abs(monthly_correlation))) %>%
  pivot_wider(names_from = season, values_from = max) %>%
  mutate(M = (spring - summer)) 

write.csv(memory, "../Compiled data/memory_3m.csv")
```

Step 3: Run driver analysis

```{r}
mem_rf <- memory %>%
  select(LakeID, var, M) %>%
  left_join(bf) %>%
  left_join(lat_long) %>%
  left_join(other_drivers) %>%
  mutate(log_SA = log(SurfaceArea_ha),
         log_max_depth = log(MaximumDepth_m),
         log_TP = log(TP_ugL_SURF),
         log_DOC = log(DOC_mgL_SURF),
         log_chla = log(Chla_ugL_SURF)) %>%
  pivot_wider(names_from = var, values_from = M) %>%
  rename(epi_temp = `Summer epi. temperature`,
         hypo_temp = `Summer bottom-water\ntemperature`,
         hypo_do = `Summer bottom-water\ndissolved oxygen`,
         vhod = `Summer VHOD`) %>%
  ungroup()

pairs(mem_rf %>% select(c("log_SA", "buoyancy_freq", 
                          "log_chla", "log_DOC", "log_TP")))
jpeg("../Figures/driver correlations.jpg", width = 5, height = 5, units = "in", res = 400)
GGally::ggpairs(mem_rf %>% 
                  select(c("log_SA", "buoyancy_freq", "log_TP")) %>%
                  rename("Log surface area (ha)" = log_SA,
                         "Buoyancy frequency (1/s)" = buoyancy_freq,
                         "Log TP (µg/L)" = log_TP))+
  theme_bw()+
  theme(text = element_text(size = 9))
dev.off()

cor(mem_rf %>% select(c("log_SA", "buoyancy_freq", "log_TP")), use = "pairwise.complete.obs")

responses <- c("epi_temp")
potential_drivers <- c("log_SA", "buoyancy_freq", 
                      "log_TP")
mem_rf_df_surf_temp <- mem_rf %>%
  select(all_of(c(responses, potential_drivers))) %>%
  na.omit() %>%
  mutate(target = "Surface-water\ntemperature")
rf1 <- randomForest(epi_temp ~ log_SA + buoyancy_freq + log_TP, 
                    data = mem_rf_df_surf_temp, 
                    ntree = 1000, 
                    importance = T)
ImpData_rf1 <- as.data.frame(importance(rf1)) %>%
  mutate(Var.Names = row.names(.)) %>%
  mutate(target = "Surface-water\ntemperature")
sa_surf_temp <- data.frame(partialPlot(rf1, as.data.frame(mem_rf_df_surf_temp), log_SA, plot = F)) %>%
    mutate(var = "Log surface\narea (km^2)")
#depth_surf_temp <- data.frame(partialPlot(rf1, as.data.frame(mem_rf_df_surf_temp), log_max_depth, plot = F)) %>%
#    mutate(var = "Log maximum\ndepth (m)")
bf_surf_temp <- data.frame(partialPlot(rf1, as.data.frame(mem_rf_df_surf_temp), buoyancy_freq, plot = F)) %>%
    mutate(var = "Buoyancy\nfrequency (1/s)")
#chla_surf_temp <- data.frame(partialPlot(rf1, as.data.frame(mem_rf_df_surf_temp), log_chla, plot = F)) %>%
#    mutate(var = "Log chlorophyll-a\n(ug/L)")
#doc_surf_temp <- data.frame(partialPlot(rf1, as.data.frame(mem_rf_df_surf_temp), log_DOC, plot = F)) %>%
#    mutate(var = "Log DOC\n(mg/L)")
tp_surf_temp <- data.frame(partialPlot(rf1, as.data.frame(mem_rf_df_surf_temp), log_TP, plot = F)) %>%
    mutate(var = "Log TP\n(µg/L)")
partials_surf_temp <- sa_surf_temp %>%
  #rbind(depth_surf_temp) %>%
  rbind(bf_surf_temp) %>%
  rbind(tp_surf_temp) %>%
  #rbind(chla_surf_temp) %>%
  #rbind(doc_surf_temp) %>%
  mutate(target = "Surface-water\ntemperature")

responses <- c("hypo_temp")
mem_rf_df_hypo_temp <- mem_rf %>%
  select(all_of(c(responses, potential_drivers))) %>%
  na.omit() %>%
  mutate(target = "Bottom-water\ntemperature")
rf2 <- randomForest(hypo_temp ~ log_SA + buoyancy_freq  + log_TP, 
                    data = mem_rf_df_hypo_temp, 
                    ntree = 1000, 
                    importance = T)
ImpData_rf2 <- as.data.frame(importance(rf2)) %>%
  mutate(Var.Names = row.names(.)) %>%
  mutate(target = "Bottom-water\ntemperature")
sa_hypo_temp <- data.frame(partialPlot(rf2, as.data.frame(mem_rf_df_hypo_temp), log_SA, plot = F)) %>%
    mutate(var = "Log surface\narea (km^2)")
#depth_hypo_temp <- data.frame(partialPlot(rf2, as.data.frame(mem_rf_df_hypo_temp), log_max_depth, plot = F)) %>%
#    mutate(var = "Log maximum\ndepth (m)")
bf_hypo_temp <- data.frame(partialPlot(rf2, as.data.frame(mem_rf_df_hypo_temp), buoyancy_freq, plot = F)) %>%
    mutate(var = "Buoyancy\nfrequency (1/s)")
#chla_hypo_temp <- data.frame(partialPlot(rf2, as.data.frame(mem_rf_df_surf_temp), log_chla, plot = F)) %>%
#    mutate(var = "Log chlorophyll-a\n(ug/L)")
#doc_hypo_temp <- data.frame(partialPlot(rf2, as.data.frame(mem_rf_df_surf_temp), log_DOC, plot = F)) %>%
#    mutate(var = "Log DOC\n(mg/L)")
tp_hypo_temp <- data.frame(partialPlot(rf2, as.data.frame(mem_rf_df_surf_temp), log_TP, plot = F)) %>%
    mutate(var = "Log TP\n(µg/L)")
partials_hypo_temp <- sa_hypo_temp %>%
  #rbind(depth_hypo_temp) %>%
  rbind(bf_hypo_temp) %>%
  #rbind(chla_hypo_temp) %>%
  #rbind(doc_hypo_temp) %>%
  rbind(tp_hypo_temp) %>%
  mutate(target = "Bottom-water\ntemperature")

responses <- c("hypo_do")
mem_rf_df_hypo_do <- mem_rf %>%
  select(all_of(c(responses, potential_drivers))) %>%
  na.omit() %>%
  mutate(target = "Bottom-water\ndissolved oxygen")
rf3 <- randomForest(hypo_do ~ log_SA + buoyancy_freq  + log_TP, 
                    data = mem_rf_df_hypo_do, 
                    ntree = 1000, 
                    importance = T)
ImpData_rf3 <- as.data.frame(importance(rf3)) %>%
  mutate(Var.Names = row.names(.)) %>%
  mutate(target = "Bottom-water\ndissolved oxygen")
sa_hypo_do <- data.frame(partialPlot(rf3, as.data.frame(mem_rf_df_hypo_do), log_SA, plot = F)) %>%
    mutate(var = "Log surface\narea (km^2)")
#depth_hypo_do <- data.frame(partialPlot(rf3, as.data.frame(mem_rf_df_hypo_do), log_max_depth, plot = F)) %>%
#    mutate(var = "Log maximum\ndepth (m)")
bf_hypo_do <- data.frame(partialPlot(rf3, as.data.frame(mem_rf_df_hypo_do), buoyancy_freq, plot = F)) %>%
    mutate(var = "Buoyancy\nfrequency (1/s)")
#chla_hypo_do <- data.frame(partialPlot(rf3, as.data.frame(mem_rf_df_surf_temp), log_chla, plot = F)) %>%
#    mutate(var = "Log chlorophyll-a\n(ug/L)")
#doc_hypo_do <- data.frame(partialPlot(rf3, as.data.frame(mem_rf_df_surf_temp), log_DOC, plot = F)) %>%
#    mutate(var = "Log DOC\n(mg/L)")
tp_hypo_do <- data.frame(partialPlot(rf3, as.data.frame(mem_rf_df_surf_temp), log_TP, plot = F)) %>%
    mutate(var = "Log TP\n(µg/L)")
partials_hypo_do <- sa_hypo_do %>%
  #rbind(depth_hypo_do) %>%
  rbind(bf_hypo_do) %>%
  rbind(tp_hypo_do) %>%
  #rbind(chla_hypo_do) %>%
  #rbind(doc_hypo_do) %>%
  mutate(target = "Bottom-water\ndissolved oxygen")

responses <- c("vhod")
mem_rf_df_vhod <- mem_rf %>%
  select(all_of(c(responses, potential_drivers))) %>%
  na.omit() %>%
  mutate(target = "Bottom-water\noxygen demand")
rf4 <- randomForest(vhod ~ log_SA + buoyancy_freq + log_TP, 
                    data = mem_rf_df_vhod, 
                    ntree = 1000, 
                    importance = T)
ImpData_rf4 <- as.data.frame(importance(rf4)) %>%
  mutate(Var.Names = row.names(.)) %>%
  mutate(target = "Bottom-water\noxygen demand")
sa_vhod <- data.frame(partialPlot(rf4, as.data.frame(mem_rf_df_vhod), log_SA, plot = F)) %>%
    mutate(var = "Log surface\narea (km^2)")
#depth_vhod <- data.frame(partialPlot(rf4, as.data.frame(mem_rf_df_vhod), log_max_depth, plot = F)) %>%
#    mutate(var = "Log maximum\ndepth (m)")
bf_vhod <- data.frame(partialPlot(rf4, as.data.frame(mem_rf_df_vhod), buoyancy_freq, plot = F)) %>%
    mutate(var = "Buoyancy\nfrequency (1/s)")
#chla_vhod <- data.frame(partialPlot(rf4, as.data.frame(mem_rf_df_surf_temp), log_chla, plot = F)) %>%
#    mutate(var = "Log chlorophyll-a\n(ug/L)")
#doc_vhod <- data.frame(partialPlot(rf4, as.data.frame(mem_rf_df_surf_temp), log_DOC, plot = F)) %>%
#    mutate(var = "Log DOC\n(mg/L)")
tp_vhod <- data.frame(partialPlot(rf4, as.data.frame(mem_rf_df_surf_temp), log_TP, plot = F)) %>%
    mutate(var = "Log TP\n(µg/L)")
partials_vhod <- sa_vhod %>%
  #rbind(depth_vhod) %>%
  rbind(bf_vhod) %>%
  #rbind(chla_vhod) %>%
  #rbind(doc_vhod) %>%
  rbind(tp_vhod) %>%
  mutate(target = "Bottom-water\noxygen demand")

rug_data <- mem_rf_df_surf_temp %>%
  full_join(mem_rf_df_hypo_temp) %>%
  full_join(mem_rf_df_vhod) %>%
  full_join(mem_rf_df_hypo_do) %>%
  pivot_longer(cols = all_of(potential_drivers),
               names_to = "var", values_to = "x") %>%
  mutate(var = recode(var, 
                      "log_SA" = "Log surface\narea (km^2)",
                      "log_max_depth" = "Log maximum\ndepth (m)",
                      "buoyancy_freq" = "Buoyancy\nfrequency (1/s)",
                      "log_DOC" = "Log DOC\n(mg/L)",
                      "log_chla" = "Log chlorophyll-a\n(ug/L)",
                      "log_TP" = "Log TP\n(µg/L)")) %>%
  mutate(target = factor(target, levels = c("Surface-water\ntemperature",
                                                "Bottom-water\ntemperature", 
                                                "Bottom-water\noxygen demand",
                                                "Bottom-water\ndissolved oxygen"))) %>%
  filter(!is.na(x))

ImpData <- ImpData_rf1 %>%
  full_join(ImpData_rf2) %>%
  full_join(ImpData_rf3) %>%
  full_join(ImpData_rf4) %>%
  mutate(target = factor(target, levels = c("Surface-water\ntemperature",
                                            "Bottom-water\ntemperature", 
                                            "Bottom-water\noxygen demand",
                                            "Bottom-water\ndissolved oxygen")))
```

Step 4: Plot identified drivers

```{r}
imp_mem <- ImpData %>%
  mutate(Var.Names = recode(Var.Names, "log_SA" = "Log surface\narea",
                            "log_max_depth" = "Log maximum\ndepth",
                            "buoyancy_freq" = "Buoyancy\nfrequency",
                            "log_DOC" = "DOC",
                            "log_chla" = "Log chlorophyll-a",
                            "log_TP" = "Log TP"),
         xlab = "") %>%
  ggplot(aes(y = `%IncMSE`, x = Var.Names)) +
  geom_col() +
  #geom_rug(sides = "", alpha = 0, outside = TRUE, length = unit(0.08, "npc")) +
  coord_cartesian(clip = "off") +
  facet_grid(target~xlab, switch = "both") + 
  labs(y = "Variable importance (% increase in MSE)") + 
  theme_bw() +
  theme(axis.title.x = element_blank(),
        strip.placement = "outside",   # format to look like title
        strip.background = element_blank(),
        strip.clip = "off",
        panel.spacing.y = unit(0.8, "lines"),
        axis.title.y = element_text(size = 9))

M_part <- partials_surf_temp %>%
  full_join(partials_hypo_temp) %>%
  full_join(partials_hypo_do) %>%
  full_join(partials_vhod)  %>%
  mutate(Type = "M")

part <- M_part %>%
  mutate(target = factor(target, levels = c("Surface-water\ntemperature",
                                            "Bottom-water\ntemperature", 
                                            "Bottom-water\noxygen demand",
                                            "Bottom-water\ndissolved oxygen"))) %>%
  group_by(Type, var, target) %>%
  arrange(x, .by_group = TRUE) %>%
  reframe(x_interp = approx(x = x, y = y, xout = seq(min(x), max(x), length.out = 10000))$x,
          y_interp = approx(x = x, y = y, xout = seq(min(x), max(x), length.out = 10000))$y) %>%
  ggplot(aes(x = x_interp)) +
  geom_hline(yintercept = 0, color = "grey60", linewidth = 0.3) +
  geom_point(aes(y = y_interp, color = y_interp), size = 0.1) +
  geom_rug(aes(x = x), sides = "b", data = rug_data, alpha = 0.4, outside = TRUE, length = unit(0.08, "npc")) +
  coord_cartesian(clip = "off") +
  facet_grid(target~var, scales = "free_x", switch = "both") +
  labs(x=NULL) +
  theme_bw() +
  scale_color_gradientn(colors = c("#c16586", "#735690", "#25469a"), 
                        limits = c(-0.3, 0.3)) +
  theme(
    strip.placement = "outside",   # format to look like title
    strip.background = element_blank(),
    #strip.text.y = element_blank(),
    strip.clip = "off",
    panel.spacing.y = unit(0.8, "lines"),
    axis.text.x = element_text(vjust=-0.5),
    axis.title.y = element_text(size = 9),
    legend.position = "none"
  ) +
  ylab("Partial dependence of SEM")

jpeg("../Figures/partials.jpg", width = 7, height = 4.5, units = "in", res = 400)
ggarrange(imp_mem, part, widths = c(0.8, 1))
dev.off()
```
