==> Data analysis/06 - Calculate trends all met limit years.Rmd <==
---
title: "Calculate trends"
author: "Abby Lewis"
date: "2023-10-10"
output: html_document
---

This file calculates trends in in-lake (temperature and dissolved oxygen) and meteorological variables. 

Table of contents:

Step 1: Load packages and data
Step 2: Calculate in-lake trends
Step 3: Calculate daily met trends


Step 1: Load packages and data

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(trend) #switched from openair 7 Nov 2023
source("sen_slope_custom.R")

# Make a directory for Figures if there isn't one
if (!dir.exists("../Figures")){dir.create("../Figures")}
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")
```

Step 2: Calculate in-lake trends

```{r}
all_data <- read.csv("../Compiled data/All_data_annual_3m.csv")

summer_bot_sen <- sen_slope_custom(all_data %>% filter(!is.na(Temp_C_BOT)), "Temp_C_BOT")
summer_surf_sen <- sen_slope_custom(all_data %>% filter(!is.na(Temp_C_SURF)), "Temp_C_SURF")
summer_surf_tp <- sen_slope_custom(all_data %>% filter(!is.na(TP_ugL_SURF)), "TP_ugL_SURF")
summer_surf_do <- sen_slope_custom(all_data %>% filter(!is.na(DO_mgL_SURF)), "DO_mgL_SURF")
summer_vhod <- sen_slope_custom(all_data %>% filter(!is.na(DO_demand_mgLd)), "DO_demand_mgLd")
summer_dif_sen <- sen_slope_custom(all_data %>% 
                                     filter(!is.na(Temp_C_BOT) & 
                                              !is.na(Temp_C_SURF)) %>%
                                     mutate(dif = Temp_C_SURF - Temp_C_BOT),
                                   "dif")
summer_dif_do <- sen_slope_custom(all_data %>% 
                                     filter(!is.na(DO_mgL_BOT) & 
                                              !is.na(DO_mgL_SURF)) %>%
                                     mutate(dif = DO_mgL_SURF - DO_mgL_BOT),
                                   "dif")
do_data <- all_data %>%
  filter(!is.na(DO_mgL_BOT)) %>% 
  group_by(LakeID) %>%
  filter(max(DO_mgL_BOT, na.rm = T) >= 1) %>%
  ungroup() 
summer_do_sen <- sen_slope_custom(do_data, "DO_mgL_BOT")

write.csv(summer_surf_sen, "../Compiled data/summer_surf_temp_trends_3m.csv", row.names = F)
write.csv(summer_surf_do, "../Compiled data/summer_surf_DO_trends_3m.csv", row.names = F)
write.csv(summer_bot_sen, "../Compiled data/summer_bot_temp_mean_trends_3m.csv", row.names = F)
write.csv(summer_do_sen, "../Compiled data/summer_bot_DO_trends_3m.csv", row.names = F)
write.csv(summer_surf_tp, "../Compiled data/summer_surf_TP_trends_3m.csv", row.names = F)
write.csv(summer_dif_sen, "../Compiled data/summer_dif_trends_3m.csv", row.names = F)
write.csv(summer_dif_do, "../Compiled data/summer_dif_DO_trends_3m.csv", row.names = F)
write.csv(summer_vhod, "../Compiled data/summer_vhod_trends_3m.csv", row.names = F)
```

Step 3: Calculate daily met trends

```{r}
#Load data
climate_era5 = read.csv("../Compiled data/historical_met_output_era5_daily.csv") %>%
  mutate(Date = as.Date(Date),
         Date_unif = ifelse(Lat > 0, Date, Date + months(6)),
         Date_unif = as.Date(Date_unif))

run_year_set <- function(climate_era5, water_trend, name){
  climate_era5_roll <- climate_era5 %>%
    mutate(doy = yday(Date_unif),
           Year = year(Date_unif)) %>%
    group_by(LakeID) %>%
    left_join(water_trend) %>%
    filter(Year >= min_year,
           Year <= max_year) %>%
    arrange(Date_unif) %>%
    mutate(Temp_C = zoo::rollmean(Temp_C, 30, align = "right", fill = NA),
           Wind_ms = zoo::rollmean(Wind_ms, 30, align = "right", fill = NA),
           Shortwave_mJm2 = zoo::rollmean(Shortwave_mJm2, 30, align = "right", fill = NA))%>%
    filter(doy < yday(as.Date("2022-08-31")),
           !is.na(Temp_C))
  
  #Calculate trends
  temp_era5_trends <- sen_slope_doys(climate_era5_roll, 
                                     "Temp_C",
                                     doys = 1:max(climate_era5_roll$doy))
  
  wind_era5_trends <- sen_slope_doys(climate_era5_roll, 
                                     "Wind_ms",
                                     doys = 1:max(climate_era5_roll$doy))
  
  light_era5_trends <- sen_slope_doys(climate_era5_roll, 
                                      "Shortwave_mJm2",
                                      doys = 1:max(climate_era5_roll$doy))
  
  #Save output
  write.csv(temp_era5_trends %>% 
              mutate(Variable = "Air_temp_C") %>%
              rbind(wind_era5_trends %>%
                      mutate(Variable = "Wind_ms")) %>% 
              rbind(light_era5_trends%>%
                      mutate(Variable = "Shortwave_mJm2")), 
            paste0("../Compiled data/climate_era5_roll_met_daily - ", name, "_years.csv"), 
            row.names = F)
}

run_year_set(climate_era5, summer_surf_sen, "summer_surf_temp")
run_year_set(climate_era5, summer_bot_sen, "summer_bot_temp")
run_year_set(climate_era5, summer_do_sen, "summer_bot_DO")
```
