==> Data analysis/07 - Plot trends all met.Rmd <==
---
title: "Air temp trends"
author: "Abby Lewis"
date: "2023-09-19"
output: html_document
---

This file plots meteorological trends and compares to in-lake trends

Table of Contents:
- Step 1: Load packages and calculated trends
- Step 2: Compare surface- vs bottom-water trends
- Step 3: Random forest for bottom-water temperature trend drivers (spring vs summer)
- Step 4: Repeat for bottom-water DO
- Step 5: Repeat for surface temperature
- Step 6: Repeat for VHOD
- Step 7: Generate plot of bottom water temperature trend drivers (spring vs summer temp)


Step 1: Load packages and calculated trends

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(trend)
library(ggh4x)
library(ggpubr)
library(randomForest)

# Make a directory for Figures if there isn't one
if (!dir.exists("../Figures")){dir.create("../Figures")}

# Load data
lat_long = read.csv("https://pasta-s.lternet.edu/package/data/eml/edi/1029/9/fadd3eaa25b5fdd1fc4efba70e660579")
met_by_bot_temp <- read_csv("../Compiled data/climate_era5_roll_met_daily - summer_bot_temp_years.csv") 
met_by_surf_temp <- read_csv("../Compiled data/climate_era5_roll_met_daily - summer_surf_temp_years.csv") 
met_by_bot_do <- read_csv("../Compiled data/climate_era5_roll_met_daily - summer_bot_DO_years.csv")
max_cor_dates <- read.csv("../Compiled data/strongest_correlation_dates_all_met_3m.csv")
mem <- read_csv("../Compiled data/memory_3m.csv") %>%
  mutate(var = recode(var, 
                      "Bottom-water temperature" = "summer_bot_mean",
                      "Bottom-water dissolved oxygen" = "summer_bot_do",
                      "Surface-water temperature" = "summer_surf_mean"))
summer_bot_mean <- read.csv("../Compiled data/summer_bot_temp_mean_trends_3m.csv") %>%
  rename(summer_bot_mean = trend) %>%
  filter(n >= 15)%>%
  dplyr::select(summer_bot_mean, LakeID) 
summer_surf_mean <- read.csv("../Compiled data/summer_surf_temp_trends_3m.csv") %>%
  rename(summer_surf_mean = trend) %>%
  filter(n >= 15)%>%
  dplyr::select(summer_surf_mean, LakeID)
summer_surf_do <- read.csv("../Compiled data/summer_surf_DO_trends_3m.csv") %>%
  rename(summer_surf_do = trend) %>%
  dplyr::select(summer_surf_do, LakeID)
summer_bot_do <- read.csv("../Compiled data/summer_bot_DO_trends_3m.csv") %>%
  rename(summer_bot_do = trend) %>%
  dplyr::select(summer_bot_do, LakeID)
summer_surf_tp <- read.csv("../Compiled data/summer_surf_tp_trends_3m.csv") %>%
  rename(summer_surf_tp = trend) %>%
  dplyr::select(summer_surf_tp, LakeID)
summer_dif <- read.csv("../Compiled data/summer_dif_trends_3m.csv") %>%
  rename(summer_dif = trend) %>%
  dplyr::select(summer_dif, LakeID)
summer_dif_do <- read.csv("../Compiled data/summer_dif_DO_trends_3m.csv") %>%
  rename(summer_dif = trend) %>%
  dplyr::select(summer_dif, LakeID)
summer_vhod <- read.csv("../Compiled data/summer_vhod_trends_3m.csv") %>%
  rename(summer_vhod = trend) %>%
  dplyr::select(summer_vhod, LakeID)
```


Step 2: Compare surface- vs bottom-water trends

```{r}
#Combine and format data
data <- summer_bot_mean %>%
  full_join(summer_surf_mean) %>%
  full_join(summer_surf_do) %>%
  full_join(summer_bot_do) %>%
  pivot_longer(cols = c(summer_bot_mean, summer_surf_mean, summer_surf_do, summer_bot_do),
               names_to = "Variable",
               values_to = "trend") %>%
  left_join(mem %>% select(LakeID, var, M) %>% distinct(), by = c("LakeID","Variable" = "var")) %>%
  mutate(layer = ifelse(str_detect(Variable, "surf"), 
                        "Surface", "Bottom"),
         var = ifelse(str_detect(Variable, "mean"), 
                      "Temperature trend\n(ºC/decade)", "DO trend\n(mg/L/decade)"),
         var = factor(var,
                      levels = c("Temperature trend\n(ºC/decade)", "DO trend\n(mg/L/decade)"))) %>%
  group_by(var, LakeID) %>%
  mutate(M = M[layer == "Bottom"]) %>%
  select(-Variable) %>%
  pivot_wider(names_from = layer,
              values_from = trend) %>%
  filter(!is.na(M))

#Calcualte correlations
cors <- data %>%
  group_by(var) %>%
  summarize(cor = cor(Surface, Bottom, method = "spearman", use = "complete.obs"),
            cor_p = cor.test(Surface, Bottom, method = "spearman", use = "complete.obs")$p.value)

#Get summary stats
min_do <- min(c(data$Bottom[data$var == "DO trend\n(mg/L/decade)"], 
                data$Surface[data$var == "DO trend\n(mg/L/decade)"]),
              na.rm = T)
max_do <- max(c(data$Bottom[data$var == "DO trend\n(mg/L/decade)"], 
                data$Surface[data$var == "DO trend\n(mg/L/decade)"]),
              na.rm = T)
min_temp <- min(c(data$Bottom[data$var == "Temperature trend\n(ºC/decade)"], 
                  data$Surface[data$var == "Temperature trend\n(ºC/decade)"]),
                na.rm = T)
max_temp <- max(c(data$Bottom[data$var == "Temperature trend\n(ºC/decade)"], 
                  data$Surface[data$var == "Temperature trend\n(ºC/decade)"]),
                na.rm = T)
mean_do <- mean(c(data$Bottom[data$var == "DO trend\n(mg/L/decade)"], 
                data$Surface[data$var == "DO trend\n(mg/L/decade)"]),
              na.rm = T)
sd_do <- sd(c(data$Bottom[data$var == "DO trend\n(mg/L/decade)"], 
                data$Surface[data$var == "DO trend\n(mg/L/decade)"]),
              na.rm = T)
mean_temp <- mean(c(data$Bottom[data$var == "Temperature trend\n(ºC/decade)"], 
                  data$Surface[data$var == "Temperature trend\n(ºC/decade)"]),
                na.rm = T)
sd_temp <- sd(c(data$Bottom[data$var == "Temperature trend\n(ºC/decade)"], 
                  data$Surface[data$var == "Temperature trend\n(ºC/decade)"]),
                na.rm = T)

#Plot surf/bot temp
a <- data %>%
  filter(var == "Temperature trend\n(ºC/decade)") %>%
  ggplot(aes(x = Bottom*10, y = Surface*10)) +
  geom_abline(color = "grey70")+
  geom_hline(yintercept = 0, color = "grey80")+
  geom_vline(xintercept = 0, color = "grey80")+
  geom_point(alpha = 0.5) +
  ylab("Surface temperature\ntrend (ºC/decade)")+
  xlab("Bottom temperature trend\n(ºC/decade)")+
  geom_label(data = cors %>%
               filter(var == "Temperature trend\n(ºC/decade)"), 
             aes(x = -10, y = 4, label =  paste0("\u03C1 = ", round(cor,2))), 
             hjust = 0, size = 3)+
  theme_bw(base_size = 10) +
  theme(axis.title = element_text(size = 9))

u <- mean(summer_dif$summer_dif, na.rm = T)
median(summer_dif$summer_dif*10, na.rm = T)
sd <- sd(summer_dif$summer_dif, na.rm = T)
p_05 <- quantile(summer_dif$summer_dif*10, 0.05, na.rm = T)
p_95 <- quantile(summer_dif$summer_dif*10, 0.95, na.rm = T)

#Plot thermal difference
b <- summer_dif %>% 
  ggplot(aes(x = summer_dif*10)) +
  geom_vline(aes(xintercept = 0), color = "grey70")+
  geom_hline(yintercept = 0, color = "grey80")+
  geom_vline(aes(xintercept = median(summer_dif*10, na.rm = T)), lty = "dotted")+
  geom_density(linewidth = 1)+
  geom_label(aes(x = -0.7, y = 0.35, 
                 label = paste0("Median = +", 
                                round(median, 2),
                                " ºC/decade")), 
             hjust = 0, size = 3,
             data = . %>% summarize(median = median(summer_dif, na.rm = T)*10))+
  xlab("Trend in thermal difference between\n surface and bottom waters (ºC/decade)")+
  ylab("\nDensity") +
  xlim(c(p_05, p_95))+
  theme_bw(base_size = 10) +
  theme(axis.title = element_text(size = 9))

dif_rf <- summer_dif %>%
  left_join(mem %>%
              filter(var == "summer_bot_mean") %>%
              select(LakeID, M)) %>%
  filter(!is.na(M), !is.na(summer_dif))

bot <- max_cor_dates %>%
  filter(var == "Summer hypo. temperature") %>%
  left_join(met_by_bot_temp, 
            by = c("LakeID", "max_doy" = "doy"), 
            relationship = "many-to-many") %>% 
  filter(Variable == "Air_temp_C",
         season == "spring") %>%
  select(LakeID, trend) %>%
  rename(spring = trend)

surf <- max_cor_dates %>%
  filter(var == "Summer epi. temperature") %>%
  left_join(met_by_surf_temp, 
            by = c("LakeID", "max_doy" = "doy"), 
            relationship = "many-to-many") %>% 
  filter(Variable == "Air_temp_C",
         season == "summer") %>%
  select(LakeID, trend) %>%
  rename(summer = trend)

#Plot categorical trends
c <- bot %>%
  left_join(surf) %>%
  left_join(dif_rf) %>%
  filter(!is.na(M)) %>%
  mutate(M_binary = ifelse(M>0, "CSEM > 0", "CSEM ≤ 0"),
         air_binary = ifelse(spring > summer, 
                             "Spring warming\nfaster than summer",
                             "Summer warming\nfaster than spring"),
         summer_binary = ifelse(summer_dif > 0,
                                "Increasing difference",
                                "Decreasing difference")) %>%
  group_by(M_binary, air_binary) %>%
  summarize(pct_inc = sum(summer_dif > 0, na.rm = T)/n()*100,
            n = n(),
            label = paste0(sum(summer_dif > 0, na.rm = T), " of\n", n)) %>%
  ggplot(aes(x = interaction(M_binary, air_binary), y = pct_inc))+
  geom_hline(yintercept = 0, color = "grey80")+
  geom_col(width = 0.6)+
  geom_text(aes(label = label,
                y = pct_inc + 6), size = 3)+
  #ylim(c(0, 100))+
  ylab("Trend in thermal difference between\nsurface and bottom > 0 (% of lakes)")+
  scale_x_discrete(guide = "axis_nested")+
  #facet_wrap(~air_binary)+
  scale_fill_manual(name = "Trend in thermal difference between\nsurface and bottom waters",
                    values = c("black", "#876BDB"))+
  theme_bw(base_size = 10)+
  theme(axis.title.x = element_blank(),
        axis.text.x = element_text(size = 8))

jpeg("../Figures/Figure 6.jpg", width = 6, height = 4, units = "in", res = 600)
ggarrange(ggarrange(a, b, 
                    heights = c(1, .6), labels = c("a", "b"), 
                    ncol = 1, nrow = 2, align = "v"), 
          ggarrange(c, 
                    ncol = 1, heights = c(1)),
          labels = c(NA, "c"), ncol = 2)
dev.off()
```


Step 3: Plot bottom temp vs spring and summer air temps

```{r}
met_trends_matched <- max_cor_dates %>%
  filter(var == "Summer hypo. temperature") %>%
  left_join(met_by_bot_temp, 
            by = c("LakeID", "max_doy" = "doy"), 
            relationship = "many-to-many") %>%
  group_by(season, var) %>%
  rename(cor_date_var = var)

df_for_rf <- summer_bot_mean %>%
  left_join(met_trends_matched %>% 
              select(LakeID, trend, season, Variable) %>%
              arrange(LakeID) %>%
              pivot_wider(names_from = Variable,
                          values_from = trend))  %>%
  pivot_wider(names_from = season,
              values_from = Air_temp_C:Shortwave_mJm2)

responses <- c("summer_bot_mean")
potential_drivers <- c("Air_temp_C_spring", "Air_temp_C_summer")
mem_rf_df_bot_temp <- df_for_rf %>%
  select(all_of(c(responses, potential_drivers))) %>%
  na.omit() %>%
  mutate(target = "Bottom-water\ntemperature")

cor_plot <- mem_rf_df_bot_temp %>%
  pivot_longer(c(Air_temp_C_spring, Air_temp_C_summer)) %>%
  mutate(name = case_match(name,
                          "Air_temp_C_spring"~"Spring",
                          "Air_temp_C_summer"~"Summer")) %>%
  group_by(name) %>%
  mutate(rho = cor.test(value, summer_bot_mean*10, method = "spearman")$estimate,
         p = cor.test(value, summer_bot_mean*10, method = "spearman")$p.value,
         p_lab = ifelse(p > 0.001, round(p, 2), p),
         p_lab = paste0("= ", round(p_lab, 3)),
         p_lab = ifelse(p_lab == "= 0",
                        "< 0.001",
                        p_lab)) %>%
  ggplot(aes(x = value*10, y = summer_bot_mean*10))+
  geom_vline(aes(xintercept = 0), color = "grey70")+
  geom_point(shape = 21, stroke = 0.3)+
  #geom_abline()+
  #geom_smooth(method = "lm")+
  geom_label(aes(label = paste0("Spearman's \u03C1 = ", round(rho, 2),
                                "\np ", p_lab), x = -2.2, y = 5, hjust= 0), 
             data = . %>% select(name, rho, p_lab) %>% distinct(), alpha = 0.5)+
  facet_grid(~name)+
  theme_bw()+
  xlab("Air temperature trend (ºC/decade)")+
  ylab("Bottom-water temperature\ntrend (ºC/decade)")

air_trends_wide_data <- mem_rf_df_bot_temp %>%
  pivot_longer(cols = all_of(potential_drivers),
               names_to = "var", values_to = "x") %>%
  filter(!is.na(x)) %>%
  mutate(var = factor(var, levels = c("Air_temp_C_spring",
                                                  "Air_temp_C_summer"),
                            labels = c("Spring", "Summer"))) 

air_trends_wide_data %>%
  group_by(var) %>%
  summarize(mean = mean(x),
            sd = sd(x))

air_trends_wide_blank <- air_trends_wide_data %>%
  mutate(spacer = "") %>%
  ggplot(aes(x = x*10))+
  geom_vline(aes(xintercept = 0), color = "grey70")+
  scale_color_manual(values = c("Spring" = "#25469a",
                               "Summer" = "#c16586"))+
  geom_density(aes(color = var))+
  facet_grid(~var) +
  theme_bw() +
  ylab("\n")+
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.clip = "off",
        strip.text.x = element_blank(),
        panel.spacing.y = unit(0.8, "lines"),
        axis.title.x = element_blank(),
        axis.line.x = element_blank(),
        axis.title.y = element_text(size = 9),
        axis.text.x = element_blank(),
        axis.text.y = element_text(color = "white"),
        axis.ticks = element_line(color = "white"),
        panel.grid = element_blank(),
        panel.border = element_blank())

jpeg("../Figures/Figure 5.jpg", res = 600, units = "in", width = 5, height = 3)
ggpubr::ggarrange(air_trends_wide_blank + theme(plot.margin = unit(c(.2,.2,0,1), 'lines')), 
                    cor_plot + theme(plot.margin = unit(c(0,.2,.2,1), 'lines')), 
                    heights = c(0.2, 1), nrow = 2)
dev.off()
```

Step 4: Repeat for bottom-water DO

```{r}
met_trends_matched <- max_cor_dates %>%
  filter(var == "Summer hypo. DO") %>%
  left_join(met_by_bot_do, 
            by = c("LakeID", "max_doy" = "doy"), 
            relationship = "many-to-many") %>%
  group_by(season, var) %>%
  rename(cor_date_var = var)

df_for_rf <- summer_bot_do %>%
  left_join(met_trends_matched %>% 
              select(LakeID, trend, season, Variable) %>%
              arrange(LakeID) %>%
              pivot_wider(names_from = Variable,
                          values_from = trend))  %>%
  pivot_wider(names_from = season,
              values_from = Air_temp_C:Shortwave_mJm2)

responses <- c("summer_bot_do")
potential_drivers <- c("Air_temp_C_spring", "Air_temp_C_summer")
mem_rf_df_bot_do <- df_for_rf %>%
  select(all_of(c(responses, potential_drivers))) %>%
  na.omit() %>%
  mutate(target = "Bottom-water\ndissolved oxygen")

cor_plot <- mem_rf_df_bot_do %>%
  pivot_longer(c(Air_temp_C_spring, Air_temp_C_summer)) %>%
  mutate(name = case_match(name,
                          "Air_temp_C_spring"~"Spring",
                          "Air_temp_C_summer"~"Summer")) %>%
  group_by(name) %>%
  mutate(rho = cor.test(value, summer_bot_do*10, method = "spearman")$estimate,
         p = cor.test(value, summer_bot_do*10, method = "spearman")$p.value,
         p_lab = ifelse(p > 0.001, round(p, 2), p),
         p_lab = paste0("= ", round(p_lab, 3)),
         p_lab = ifelse(p_lab == "= 0",
                        "< 0.001",
                        p_lab)) %>%
  ggplot(aes(x = value*10, y = summer_bot_do*10))+
  geom_vline(aes(xintercept = 0), color = "grey70")+
  geom_point(shape = 21, stroke = 0.3)+
  #geom_abline()+
  #geom_smooth(method = "lm")+
  geom_label(aes(label = paste0("Spearman's \u03C1 = ", round(rho, 2),
                                "\np ", p_lab), x = -1.7, y = 3.5, hjust= 0), 
             data = . %>% select(name, rho, p_lab) %>% distinct(), alpha = 0.5)+
  facet_grid(~name)+
  theme_bw()+
  xlab("Air temperature trend (ºC/decade)")+
  ylab("Bottom-water dissolved\noxygen trend (ºC/decade)")

air_trends_wide_data <- mem_rf_df_bot_do %>%
  pivot_longer(cols = all_of(potential_drivers),
               names_to = "var", values_to = "x") %>%
  filter(!is.na(x)) %>%
  mutate(var = factor(var, levels = c("Air_temp_C_spring",
                                                  "Air_temp_C_summer"),
                            labels = c("Spring", "Summer"))) 

air_trends_wide_data %>%
  group_by(var) %>%
  summarize(mean = mean(x),
            sd = sd(x))

air_trends_wide_blank <- air_trends_wide_data %>%
  mutate(spacer = "") %>%
  ggplot(aes(x = x*10))+
  geom_vline(aes(xintercept = 0), color = "grey70")+
  scale_color_manual(values = c("Spring" = "#25469a",
                               "Summer" = "#c16586"))+
  geom_density(aes(color = var))+
  facet_grid(~var) +
  theme_bw() +
  ylab("\n")+
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.clip = "off",
        strip.text.x = element_blank(),
        panel.spacing.y = unit(0.8, "lines"),
        axis.title.x = element_blank(),
        axis.line.x = element_blank(),
        axis.title.y = element_text(size = 9),
        axis.text.x = element_blank(),
        axis.text.y = element_text(color = "white"),
        axis.ticks = element_line(color = "white"),
        panel.grid = element_blank(),
        panel.border = element_blank())

jpeg("../Figures/simple_Fig5_DO.jpg", res = 300, units = "in", width = 5, height = 3)
ggpubr::ggarrange(air_trends_wide_blank + theme(plot.margin = unit(c(.2,.2,0,1), 'lines')), 
                    cor_plot + theme(plot.margin = unit(c(0,.2,.2,1), 'lines')), 
                    heights = c(0.2, 1), nrow = 2)
dev.off()
```

Step 5: Repeat for surface temperature

```{r}
met_trends_matched <- max_cor_dates %>%
  filter(var == "Summer epi. temperature") %>%
  left_join(met_by_bot_do, 
            by = c("LakeID", "max_doy" = "doy"), 
            relationship = "many-to-many") %>%
  group_by(season, var) %>%
  rename(cor_date_var = var)

df_for_rf <- summer_surf_mean %>%
  left_join(met_trends_matched %>% 
              select(LakeID, trend, season, Variable) %>%
              arrange(LakeID) %>%
              pivot_wider(names_from = Variable,
                          values_from = trend))  %>%
  pivot_wider(names_from = season,
              values_from = Air_temp_C:Shortwave_mJm2)

responses <- c("summer_surf_mean")
potential_drivers <- c("Air_temp_C_spring", "Air_temp_C_summer")
mem_rf_df_surf_temp <- df_for_rf %>%
  select(all_of(c(responses, potential_drivers))) %>%
  na.omit() %>%
  mutate(target = "Surface-water\ntemperature")

cor_plot <- mem_rf_df_surf_temp %>%
  pivot_longer(c(Air_temp_C_spring, Air_temp_C_summer)) %>%
  mutate(name = case_match(name,
                          "Air_temp_C_spring"~"Spring",
                          "Air_temp_C_summer"~"Summer")) %>%
  group_by(name) %>%
  mutate(rho = cor.test(value, summer_surf_mean*10, method = "spearman")$estimate,
         p = cor.test(value, summer_surf_mean*10, method = "spearman")$p.value,
         p_lab = ifelse(p > 0.001, round(p, 2), p),
         p_lab = paste0("= ", round(p_lab, 3)),
         p_lab = ifelse(p_lab == "= 0",
                        "< 0.001",
                        p_lab)) %>%
  ggplot(aes(x = value*10, y = summer_surf_mean*10))+
  geom_vline(aes(xintercept = 0), color = "grey70")+
  geom_point(shape = 21, stroke = 0.3)+
  #geom_abline()+
  #geom_smooth(method = "lm")+
  geom_label(aes(label = paste0("Spearman's \u03C1 = ", round(rho, 2),
                                "\np ", p_lab), x = -3, y = -2.8, hjust= 0), 
             data = . %>% select(name, rho, p_lab) %>% distinct(), alpha = 0.5)+
  facet_grid(~name)+
  theme_bw()+
  xlab("Air temperature trend (ºC/decade)")+
  ylab("Surface-water temperature\ntrend (ºC/decade)")

air_trends_wide_data <- mem_rf_df_surf_temp %>%
  pivot_longer(cols = all_of(potential_drivers),
               names_to = "var", values_to = "x") %>%
  filter(!is.na(x)) %>%
  mutate(var = factor(var, levels = c("Air_temp_C_spring",
                                                  "Air_temp_C_summer"),
                            labels = c("Spring", "Summer"))) 

air_trends_wide_data %>%
  group_by(var) %>%
  summarize(mean = mean(x),
            sd = sd(x))

air_trends_wide_blank <- air_trends_wide_data %>%
  mutate(spacer = "") %>%
  ggplot(aes(x = x*10))+
  geom_vline(aes(xintercept = 0), color = "grey70")+
  scale_color_manual(values = c("Spring" = "#25469a",
                               "Summer" = "#c16586"))+
  geom_density(aes(color = var))+
  facet_grid(~var) +
  theme_bw() +
  ylab("\n")+
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.clip = "off",
        strip.text.x = element_blank(),
        panel.spacing.y = unit(0.8, "lines"),
        axis.title.x = element_blank(),
        axis.line.x = element_blank(),
        axis.title.y = element_text(size = 9),
        axis.text.x = element_blank(),
        axis.text.y = element_text(color = "white"),
        axis.ticks = element_line(color = "white"),
        panel.grid = element_blank(),
        panel.border = element_blank())

jpeg("../Figures/simple_Fig5_surf.jpg", res = 300, units = "in", width = 5, height = 3)
ggpubr::ggarrange(air_trends_wide_blank + theme(plot.margin = unit(c(.2,.2,0,1), 'lines')), 
                    cor_plot + theme(plot.margin = unit(c(0,.2,.2,1), 'lines')), 
                    heights = c(0.2, 1), nrow = 2)
dev.off()
```

Step 6: Repeat for VHOD

```{r}
met_trends_matched <- max_cor_dates %>%
  filter(var == "Summer VHOD") %>%
  left_join(met_by_bot_do, 
            by = c("LakeID", "max_doy" = "doy"), 
            relationship = "many-to-many") %>%
  group_by(season, var) %>%
  rename(cor_date_var = var)

df_for_rf <- summer_vhod %>%
  left_join(met_trends_matched %>% 
              select(LakeID, trend, season, Variable) %>%
              arrange(LakeID) %>%
              pivot_wider(names_from = Variable,
                          values_from = trend))  %>%
  pivot_wider(names_from = season,
              values_from = Air_temp_C:Shortwave_mJm2)

responses <- c("summer_vhod")
potential_drivers <- c("Air_temp_C_spring", "Air_temp_C_summer")
mem_rf_df_vhod <- df_for_rf %>%
  select(all_of(c(responses, potential_drivers))) %>%
  na.omit() %>%
  mutate(target = "Bottom-water\noxygen demand")

cor_plot <- mem_rf_df_vhod %>%
  pivot_longer(c(Air_temp_C_spring, Air_temp_C_summer)) %>%
  mutate(name = case_match(name,
                          "Air_temp_C_spring"~"Spring",
                          "Air_temp_C_summer"~"Summer")) %>%
  group_by(name) %>%
  mutate(rho = cor.test(value, summer_vhod*10, method = "spearman")$estimate,
         p = cor.test(value, summer_vhod*10, method = "spearman")$p.value,
         p_lab = ifelse(p > 0.001, round(p, 2), p),
         p_lab = paste0("= ", round(p_lab, 3)),
         p_lab = ifelse(p_lab == "= 0",
                        "< 0.001",
                        p_lab)) %>%
  ggplot(aes(x = value*10, y = summer_vhod*10))+
  geom_vline(aes(xintercept = 0), color = "grey70")+
  geom_point(shape = 21, stroke = 0.3)+
  #geom_abline()+
  #geom_smooth(method = "lm")+
  geom_label(aes(label = paste0("Spearman's \u03C1 = ", round(rho, 2),
                                "\np ", p_lab), x = -3, y = 0.15, hjust= 0), 
             data = . %>% select(name, rho, p_lab) %>% distinct(), alpha = 0.5)+
  facet_grid(~name)+
  theme_bw()+
  xlab("Air temperature trend (ºC/decade)")+
  ylab("Bottom-water oxygen\ndemand trend (ºC/decade)")

air_trends_wide_data <- mem_rf_df_vhod %>%
  pivot_longer(cols = all_of(potential_drivers),
               names_to = "var", values_to = "x") %>%
  filter(!is.na(x)) %>%
  mutate(var = factor(var, levels = c("Air_temp_C_spring",
                                                  "Air_temp_C_summer"),
                            labels = c("Spring", "Summer"))) 

air_trends_wide_data %>%
  group_by(var) %>%
  summarize(mean = mean(x),
            sd = sd(x))

air_trends_wide_blank <- air_trends_wide_data %>%
  mutate(spacer = "") %>%
  ggplot(aes(x = x*10))+
  geom_vline(aes(xintercept = 0), color = "grey70")+
  scale_color_manual(values = c("Spring" = "#25469a",
                               "Summer" = "#c16586"))+
  geom_density(aes(color = var))+
  facet_grid(~var) +
  theme_bw() +
  ylab("\n")+
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.clip = "off",
        strip.text.x = element_blank(),
        panel.spacing.y = unit(0.8, "lines"),
        axis.title.x = element_blank(),
        axis.line.x = element_blank(),
        axis.title.y = element_text(size = 9),
        axis.text.x = element_blank(),
        axis.text.y = element_text(color = "white"),
        axis.ticks = element_line(color = "white"),
        panel.grid = element_blank(),
        panel.border = element_blank())

jpeg("../Figures/simple_Fig5_vhod.jpg", res = 300, units = "in", width = 5, height = 3)
ggpubr::ggarrange(air_trends_wide_blank + theme(plot.margin = unit(c(.2,.2,0,1), 'lines')), 
                    cor_plot + theme(plot.margin = unit(c(0,.2,.2,1), 'lines')), 
                    heights = c(0.2, 1), nrow = 2)
dev.off()
```

Step 7: Generate plot of bottom water temperature trend drivers (spring vs summer temp)

```{r}
p <- dif_rf %>%
  left_join(lat_long)

with_temp = read.csv("../Compiled data/All_data_annual_3m.csv")
other_drivers <- with_temp %>%
  group_by(LakeID) %>%
  select(DO_mgL_BOT, Temp_C_BOT, buoyancy_freq,LakeID) %>%
  summarize(across(everything(), \(x) median(x, na.rm = TRUE)))

jpeg("../Figures/Thermal difference by bottom temp.jpg", res = 300, width = 4, height = 4, units = "in")
p %>%
  left_join(surf) %>%
  left_join(bot) %>%
  left_join(mem %>%
              filter(var == "summer_bot_mean") %>%
              select(LakeID, M)) %>%
  mutate(M_binary = ifelse(M > 0,
                           "CSEM > 0",
                           "CSEM ≤ 0"),
         M_binary = factor(M_binary,
                           levels = c("CSEM > 0", "CSEM ≤ 0")),
         air_dif = ifelse(summer > spring, 
                          "Faster summer warming",
                          "Faster spring warming")) %>%
  left_join(other_drivers) %>%
  ggplot(aes(x = Temp_C_BOT, y = summer_dif*10, color = MaximumDepth_m))+
  geom_hline(yintercept = 0, color = "grey70")+
  geom_point()+
  #geom_smooth()+
  scale_color_viridis_c(trans = "log10", direction = -1, name = "Maximum depth (m)")+
  theme_bw()+
  xlab("Median bottom-water temperature")+
  ylab("Trend in thermal difference between\nsurface and bottom waters (ºC/decade)")+
  theme(legend.position = "bottom")+
  facet_grid(M_binary~air_dif)
dev.off()
```
