==> Data analysis/08 - ISIMIP_correlations.Rmd <==
---
title: "ISIMIP analysis"
author: "Abby Lewis"
date: "2024-01-19"
output: html_document
---

This file calculates ecological memory in ISIMIP lakes and relates air temperature dates to stratification onset

Table of Contents:
- Step 1: Load packages and data
- Step 2: Calculate rolling correlations
- Step 3: Calculate and plot memory
- Step 4: Plots rolling means
- Step 5: Plot dates vs. stratification


Step 1: Load packages and data

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ncdf4)
library(tidyverse)
library(lubridate)
library(rLakeAnalyzer)
source("../Data processing/thermo.depth.density.R")
source("correlations_doy.R")

all <- read.csv("../Compiled data/Correlations - 30 day rolling mean - surf bot.csv")
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")

all_export <- all %>%
  filter(grepl("temp", var)) %>%
  dplyr::select(LakeID) %>%
  unique() %>%
  left_join(lat_long) %>%
  dplyr::select(hylak_id, LakeName, LakeID, Latitude_DD, Longitude_DD) %>%
  filter(!is.na(hylak_id))

isimip <- nc_open('../External data/hydrolakes_id.nc') 
lakes <- ncvar_get(isimip,"id")
lat <- ncvar_get(isimip,"lat")
lon <- ncvar_get(isimip,"lon")
combined_df = data.frame(ID = as.vector(lakes),
                           lon = rep(lon,length(lat)),
                           lat = rep(lat,each=length(lon))
) %>%
  filter(!is.na(ID))
all_export2 <- all_export %>% 
  filter(hylak_id %in% combined_df$ID)
write.csv(all_export2, "../Data processing/lakes_for_isimip.csv", row.names = F)

lakes <- all_export2
climate <- read.csv("../Compiled data/ISIMIP_ERA5_temp_1981_2005.csv")%>%
  left_join(lat_long)

summer <- read.csv("../Compiled data/ISIMIP_summer_avgs_v3 - surf bot.csv")
```

Step 2: Calculate rolling correlations

```{r}
daily_temp <- climate %>%
  mutate(Date = as.Date(Date),
         Date_unif = ifelse(Lat < 0, Date + months(6), Date),
         Date_unif = as.Date(Date_unif, origin = "1970-01-01")) %>%
  dplyr::select(-Lat, -Lon) %>%
  mutate(Year = year(Date_unif))

many_lake_temp_sum <- summer %>%
  left_join(daily_temp) %>%
  mutate(doy = yday(Date_unif),
         Year = ifelse(month(Date_unif) >= 9, Year + 1, Year)) %>%
  group_by(LakeID) %>%
  arrange(Date_unif) %>%
  mutate(Temp_C = zoo::rollmean(Temp_C, 30, align = "right", fill = NA))
# ---

epi_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_SURF", "Temp_C", 
                                    "Summer epi. temperature ")

hypo_temp_roll <- correlations_doy(many_lake_temp_sum, 
                                    "Temp_C_BOT", "Temp_C", 
                                    "Summer hypo. temperature ")

all <- epi_temp_roll %>%
  mutate(var = "Summer epi. temperature") %>%
  full_join(hypo_temp_roll %>%
              mutate(var = "Summer hypo. temperature")) 

write.csv(all, "../Compiled data/ISIMIP correlations - 30 day rolling mean v3.csv", row.names = F)
```

Step 3: Calculate and plot memory

```{r}
alpha <- 0.05
all <- read.csv("../Compiled data/ISIMIP correlations - 30 day rolling mean v3.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F))

sp <- yday(c("2022-03-01", "2022-05-31"))
su <- yday(c("2022-07-01", "2022-08-31"))

memory <- all %>% 
  filter(var %in% c("Summer epi. temperature",
                    "Summer hypo. temperature")) %>%
  mutate(var = factor(var, levels = c("Summer epi. temperature",
                                      "Summer hypo. temperature"),
                      labels = c("Summer epi.\ntemperature",
                                 "Summer hypo.\ntemperature")),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  group_by(var) %>%
  mutate(p = wilcox.test(monthly_correlation)$p.value) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season, p) %>%
  summarize(max = abs(max(monthly_correlation))) %>%
  pivot_wider(names_from = season, values_from = max) %>%
  mutate(M = (spring - summer)) 

jpeg("../Figures/ISIMIP monthly correlations - sum long concept v3.jpg", width = 3.5, height = 2, res = 600, units = "in")
memory %>%
  group_by(var) %>% 
  mutate(n = length(unique(LakeID)),
         var = paste0(var, " (n = ",n,")")) %>%
  mutate(var = factor(var, 
                      levels = c("Summer epi.\ntemperature (n = 43)",
                                 "Summer hypo.\ntemperature (n = 43)"),
                      labels = c("Surface-water\ntemperature (n = 43)",
                                 "Bottom-water\ntemperature (n = 43)")))%>%
  group_by(var) %>%
  mutate(color = ifelse(p > alpha, "Insignificant", 
                        ifelse(median(M) > 0, "median > 0", "median < 0")),
         color = factor(color, levels = c("median < 0", "Insignificant", "median > 0"))) %>%
  ggplot(aes(y = var, x = M, color = color))+
  geom_vline(xintercept = 0, color = "grey70")+
  geom_boxplot()+
  annotate("text", x = 0.7, y = 3, label = "More correlated\nwith spring", 
           lineheight = .9, size = 3, hjust = 0, color = "#25469a")+
  annotate("text", x = -0.7, y = 3, label = "More correlated\nwith summer", 
           lineheight = .9, size = 3, hjust = 1, color = "#c16586")+
  theme_bw() +
  ggtitle("\n")+
  scale_y_discrete(limits=rev, position = "right") +
  scale_x_reverse()+
  xlab("Relative correlation with spring\nvs. summer air temperature")+
  guides(color = guide_legend(nrow = 2))+
  theme(axis.title.y = element_blank(),
        axis.title.x = element_text(size= 8),
        legend.position = "none",
        legend.text = element_text(size = 8),
        legend.key.size = unit(1, "line"),
        legend.title = element_text(size = 8),
        legend.margin=ggplot2::margin(c(0,0,0,0)),
        legend.box.margin=ggplot2::margin(t = -2,l = 0, r = 0, b = -5),
        #axis.text.y=element_text(colour=c("#25469a","#25469a","#25469a","#c16586"))
        )+
  scale_color_manual(values = c("#c16586", "#25469a"), name = paste0("Boxplot color\n(p < ", alpha, ")"))+
  coord_cartesian(xlim = c(0.65,-0.65), ylim = c(1,2), clip = "off")
dev.off()
```

Step 5: Plot dates vs. stratification

```{r}
dates <- all %>% 
  filter(var %in% c("Summer epi. temperature",
                    "Summer hypo. temperature")) %>%
  mutate(var = factor(var, levels = c("Summer epi. temperature",
                                      "Summer hypo. temperature"),
                      labels = c("Summer\nepi. temperature",
                                 "Summer\nhypo. temperature")),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  group_by(var) %>%
  mutate(p = wilcox.test(monthly_correlation)$p.value) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season, p) %>%
  summarize(max_day = doy[which.max(monthly_correlation)],
            max = abs(max(monthly_correlation))) %>%
  pivot_wider(names_from = season, values_from = c(max, max_day)) 

strat_onset <- read.csv("../Compiled data/ISIMIP_strat_onset_v3_yearly.csv")

dates_cor <- dates %>%
  filter(var == "Summer\nhypo. temperature") %>%
  left_join(lat_long %>% select(LakeID, hylak_id)) %>%
  left_join(strat_onset %>%
              mutate(strat_onset = yday(as.Date(Date_unif)))) %>%
  group_by(LakeID, hylak_id, max_spring, max_summer, max_day_spring) %>%
  summarize(mean = mean(strat_onset),
            sd = sd(strat_onset))

jpeg("../Figures/Correlation between stratification onset and hypo. temp.jpg", width = 5.5, height = 4, res = 300, units = "in")
dates_cor %>%
  mutate(Memory = max_spring - max_summer) %>%
  ggplot(aes(x = mean, y = max_day_spring)) +
  geom_abline(slope = 1) +
  geom_smooth(method = "lm", color = "grey50", lty = "11")+
  geom_errorbar(aes(xmin = mean-sd, xmax = mean+sd))+
  geom_point(aes(fill = Memory, group = LakeID), size = 2, shape = 21, color = "black") +
  scale_fill_gradient2(low = "#c16586", mid = "white", high = "#25469a", 
                       midpoint = 0, limits = c(-0.7, 0.7), 
                       breaks = c(-0.6, -0.3, 0, 0.3, 0.6),
                       labels = c("More correlated\nwith summer",
                                  "-0.3",
                                  "0",
                                  "0.3",
                                  "More correlated\nwith spring"),
                       name = "")+
  ylab("Day of year most correlated with\nsummer bottom-water temperature") +
  xlab("Stratification onset (day of year)")+
  theme_bw()
dev.off()

summary(lm(max_day_spring ~ mean, data = dates_cor))
```
