==> Data analysis/08 - ISIMIP_correlations_split.Rmd <==
---
title: "ISIMIP analysis"
author: "Abby Lewis"
date: "2024-01-19"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ncdf4)
library(tidyverse)
library(lubridate)
library(rLakeAnalyzer)
source("../Data processing/thermo.depth.density.R")
source("correlations_doy.R")

all <- read.csv("../Compiled data/Correlations - 30 day rolling mean.csv")
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")

all_export <- all %>%
  filter(grepl("temp", var)) %>%
  dplyr::select(LakeID) %>%
  unique() %>%
  left_join(lat_long) %>%
  dplyr::select(hylak_id, LakeName, LakeID, Latitude_DD, Longitude_DD) %>%
  filter(!is.na(hylak_id))

isimip <- nc_open('../External data/hydrolakes_id.nc') 
lakes <- ncvar_get(isimip,"id")
lat <- ncvar_get(isimip,"lat")
lon <- ncvar_get(isimip,"lon")
combined_df = data.frame(ID = as.vector(lakes),
                           lon = rep(lon,length(lat)),
                           lat = rep(lat,each=length(lon))
) %>%
  filter(!is.na(ID))
all_export2 <- all_export %>% 
  filter(hylak_id %in% combined_df$ID)
write.csv(all_export2, "../Data processing/lakes_for_isimip.csv", row.names = F)

lakes <- all_export2
climate <- read.csv("../Compiled data/ISIMIP_ERA5_temp_1981_2005.csv")%>%
  left_join(lat_long)

summer <- read.csv("../Compiled data/ISIMIP_summer_avgs_v3.csv") %>%
  pivot_wider(names_from = "Layer", values_from = "Temp_C", names_prefix = "Temp_C_")

```

```{r}
max(summer$Year)
min(summer$Year)
max(climate$Date)
min(climate$Date)

daily_temp <- climate %>%
  mutate(Date = as.Date(Date),
         Date = ifelse(Lat < 0, Date - months(6), Date),
         Date = as.Date(Date, origin = "1970-01-01")) %>%
  dplyr::select(-Lat, -Lon) %>%
  mutate(Year = year(Date))

many_lake_temp_sum <- summer %>%
  left_join(daily_temp) %>%
  mutate(doy = yday(Date),
         Year = ifelse(month(Date) >= 9, Year + 1, Year),
         era = dplyr::if_else(Date <= as.Date("2000-01-01"), 
                      "1980-2000",
                      "2001-2021")) %>%
  group_by(LakeID) %>%
  arrange(Date) %>%
  mutate(Temp_C = zoo::rollmean(Temp_C, 30, align = "right", fill = NA))
# ---

epi_temp_roll_1980_2000 <- correlations_doy(many_lake_temp_sum %>% filter(era == "1980-2000"), 
                                    "Temp_C_EPI", "Temp_C", 
                                    "Summer epi. temperature ")


epi_temp_roll_2001_2021 <- correlations_doy(many_lake_temp_sum %>% filter(era == "2001-2021"), 
                                    "Temp_C_EPI", "Temp_C", 
                                    "Summer epi. temperature ")

hypo_temp_roll_1980_2000 <- correlations_doy(many_lake_temp_sum %>% filter(era == "1980-2000"), 
                                    "Temp_C_HYPO", "Temp_C", 
                                    "Summer hypo. temperature ")

hypo_temp_roll_2001_2021 <- correlations_doy(many_lake_temp_sum %>% filter(era == "2001-2021"),
                                    "Temp_C_HYPO", "Temp_C", 
                                    "Summer hypo. temperature ")

all <- epi_temp_roll_1980_2000 %>%
  mutate(var = "Summer epi. temperature", 
         era = "1980-2000") %>%
  rbind(epi_temp_roll_2001_2021 %>%
          mutate(var = "Summer epi. temperature",
                 era = "2001-2021")) %>%
  rbind(hypo_temp_roll_1980_2000 %>%
          mutate(var = "Summer hypo. temperature",
                 era = "1980-2000")) %>%
  rbind(hypo_temp_roll_2001_2021 %>%
          mutate(var = "Summer hypo. temperature", 
                 era = "2001-2021"))

write.csv(all, "../Compiled data/ISIMIP correlations - eras.csv", row.names = F)
```

```{r}
alpha <- 0.05
all <- read.csv("../Compiled data/ISIMIP correlations - eras.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F))

sp <- yday(c("2022-03-01", "2022-05-31"))
su <- yday(c("2022-07-01", "2022-08-31"))

memory <- all %>% 
  filter(var %in% c("Summer epi. temperature",
                    "Summer hypo. temperature")) %>%
  mutate(var = factor(var, levels = c("Summer epi. temperature",
                                      "Summer hypo. temperature"),
                      labels = c("Summer epi.\ntemperature",
                                 "Summer hypo.\ntemperature")),
         season = ifelse(doy >= sp[1] & doy <= sp[2], "spring", 
                         ifelse(doy >= su[1] & doy <= su[2], "summer", NA))) %>%
  group_by(var, era) %>%
  mutate(p = wilcox.test(monthly_correlation)$p.value) %>%
  filter(!is.na(season)) %>%
  group_by(LakeID, var, season, p, era) %>%
  summarize(max = max(abs(monthly_correlation))) %>%
  pivot_wider(names_from = season, values_from = max) %>%
  mutate(M = (summer - spring)) 

jpeg("../Figures/ISIMIP monthly correlations - eras.jpg", width = 3.5, height = 2, res = 600, units = "in")
memory %>%
  group_by(var, era) %>% 
  mutate(var = factor(var, 
                      levels = c("Summer epi.\ntemperature",
                                 "Summer hypo.\ntemperature"),
                      labels = c("Surface-water\ntemperature",
                                 "Bottom-water\ntemperature")))%>%
  group_by(var, era) %>%
  mutate(color = ifelse(p > alpha, "Insignificant", 
                        ifelse(median(M) > 0, "median > 0", "median < 0")),
         color = factor(color, levels = c("median < 0", "Insignificant", "median > 0")),
         color = paste0(color, "; ", era)) %>%
  ggplot(aes(y = var, x = M, color = color))+
  geom_vline(xintercept = 0, color = "grey70")+
  geom_boxplot()+
  annotate("text", x = -0.7, y = 3, label = "More correlated\nwith spring", 
           lineheight = .9, size = 3, hjust = 0, color = "#25469a")+
  annotate("text", x = 0.7, y = 3, label = "More correlated\nwith summer", 
           lineheight = .9, size = 3, hjust = 1, color = "#c16586")+
  theme_bw() +
  ggtitle("\n")+
  scale_y_discrete(limits=rev, position = "right") +
  xlab("Relative correlation with summer\nvs. spring air temperature")+
  guides(color = guide_legend(nrow = 2))+
  theme(axis.title.y = element_blank(),
        axis.title.x = element_text(size= 8),
        legend.text = element_text(size = 8),
        legend.key.size = unit(1, "line"),
        legend.title = element_text(size = 8),
        legend.margin=margin(c(0,0,0,0)),
        legend.box.margin=margin(t = -2,l = 0, r = 0, b = -5),
        #axis.text.y=element_text(colour=c("#25469a","#25469a","#25469a","#c16586"))
        )+
  scale_color_manual(values = c(colorspace::lighten("#25469a",.5), "#25469a",
                                colorspace::lighten("#c16586", .5), "#c16586"), 
                     name = paste0("Boxplot color\n(p < ", alpha, ")"))+
  coord_cartesian(xlim = c(-0.65,0.65), ylim = c(1,2), clip = "off")
dev.off()
```

Plots of rolling means

```{r}
alpha <- 0.001
all <- read.csv("../Compiled data/ISIMIP correlations - eras.csv") %>%
  mutate(sig = ifelse(p < alpha, T, F))

group_fun <- function(sig) {
  group <- rep(1, length(sig))
  for (i in 2:length(sig)) {
    if(sig[i] == sig[i - 1]) {
      group[i] <- group[i - 1]
    } else {
      group[i] <- group[i - 1] + 1
    }
  }
  return(group)
}

all_plot <- all %>%
  group_by(var, doy, sig, era) %>%
  summarise(mean = mean(monthly_correlation, na.rm = T),
            sd = sd(monthly_correlation, na.rm = T)) 

all_plot$group <- 1
for(i in 1:length(unique(all_plot$var))){
  for(j in 1:length(unique(all_plot$era))){
    all_plot$group[all_plot$var == unique(all_plot$var)[i] & 
                   all_plot$era == unique(all_plot$era)[j]] <- 
      group_fun(all_plot$sig[all_plot$var == unique(all_plot$var)[i] & 
                             all_plot$era == unique(all_plot$era)[j]])
  }
}

all_plot1 <- all_plot %>%
  mutate(DateTime = as.Date("2022-01-01") + days(doy))

all_plot2 <- all_plot %>%
  group_by(era) %>%
  arrange(var) %>%
  mutate(DateTime = as.Date("2022-01-01") + days(doy) + hours(23) + minutes(59),
         mean = ifelse(!is.na(lead(var)) & lead(var) == var, lead(mean), mean), #to get smooth lines
         sd = ifelse(lead(var) == var, lead(sd), sd))

colored_sd <- all_plot1 %>%
  full_join(all_plot2) %>%
  filter(var %in% c("Summer epi. temperature",
                    "Summer hypo. temperature")) %>%
  mutate(var = factor(var, 
                      levels = c("Summer epi. temperature",
                                 "Summer hypo. temperature"),
                      labels = c("Summer epi. temperature",
                                 "Summer hypo. temperature")),
         layer = ifelse(grepl("epi", var), "epi", "hypo")) %>%
  ggplot() + 
  geom_hline(yintercept=0)+
  geom_vline(xintercept = as_datetime("2022-04-21"), color = "grey50", lty = "dashed")+
  geom_ribbon(aes(x = DateTime, 
                  ymin = mean - sd, ymax = mean + sd,
                  fill = sig, group = group),
              alpha = 0.5) +
  geom_line(aes(x = DateTime, y = mean)) +
  xlab("Air temperature date \n(right-aligned 30-day rolling mean)")+
  ylab("Correlation")+
  theme_bw()+
  scale_fill_manual(values = c("grey40","#0FA9E6"))+
  labs(fill = paste0("p < ",alpha)) +
  facet_grid(rows = vars(var), cols = vars(era)) +
  theme(legend.position = "bottom",
        strip.background = element_blank())

jpeg("../Figures/Monthly correlations - all daily roll eras.jpg", width = 8, height = 5, res = 300, units = "in")
colored_sd
dev.off()

length(unique(all$LakeID[!is.na(all$monthly_correlation)]))
```
